refactor(sync): fushi_sync_server 按域拆成 7 个 part（B3）
FushiSyncServer 3194 行主体拆成主库 + fushi_sync_server/*.part.dart：
auth / pairing / lookup / library / video / sync_state / webdav，每个 part
是一个 `extension _FushiSyncServerX on FushiSyncServer`，方法逐字搬移、
零行为改动。主库只留值类型、构造/字段、start/stop、_gzipTextMiddleware、
_handleRequest 路由表、公开回调与计数 getter、ByteRange / ExportPackageCache。

extension 体内看不到宿主类的 static，所以 private static helper 提到库顶层
（跨域共享的 5 个放主库：_constantTimeEquals / _extractVideoId /
_decodeHeaderValue / _coverFile / _guessContentType；只被一个域用的放对应
part 尾部）。公开 static generateToken 在 pairing.part 里改为类名限定调用。
对包外零可见性变化。

守卫随扫描面搬家：新增 test/sync/fushi_sync_server_source_corpus.dart
（复用 helpers/part_corpus.dart：主库 + 磁盘枚举全部 part）；hardening /
asset_gate / interconnect_profile_transfer / interconnect_token_display /
pin_no_plaintext 五条改读合并语料，hardening 里 _serializeDavWrite 切片终点
从 _handlePair（已在别的 part、语料里排在前面）换成同 part 的 _handlePropfind；
remote_lookup_routes_test 的负向清单改为按来源函数取语料（只扫主库会让搬进
lookup.part 的重复壳真空通过）。

验证：flutter analyze 全量零问题；flutter test test/sync test/tools 2882 全过。
