+++ fushi/lib/src/sync/fushi_sync_server.dart
@@ -29,6 +29,14 @@ import 'package:path/path.dart' as p;
 import 'package:shelf/shelf.dart' as shelf;
 import 'package:shelf/shelf_io.dart' as shelf_io;
 
+part 'fushi_sync_server/auth.part.dart';
+part 'fushi_sync_server/pairing.part.dart';
+part 'fushi_sync_server/lookup.part.dart';
+part 'fushi_sync_server/library.part.dart';
+part 'fushi_sync_server/video.part.dart';
+part 'fushi_sync_server/sync_state.part.dart';
+part 'fushi_sync_server/webdav.part.dart';
+
 /// Embedded WebDAV-style server used for device-to-device LAN sync.
 ///
 /// SECURITY (HBK-AUDIT-011): transport is plain HTTP and auth is HTTP Basic,
@@ -134,6 +142,55 @@ const List<String> _kDictionaryMediaTokenParams = <String>[
   'yomitan_api_key',
 ];
 
+// ── 跨域共享的顶层 helper（原 FushiSyncServer 的 private static；extension 体内看不到宿主类
+//    的 static，故提到库顶层——对包外零可见性变化）。
+
+bool _constantTimeEquals(Uint8List a, Uint8List b) {
+  final len = a.length > b.length ? a.length : b.length;
+  var result = a.length ^ b.length;
+  for (var i = 0; i < len; i++) {
+    result |= (i < a.length ? a[i] : 0) ^ (i < b.length ? b[i] : 0);
+  }
+  return result == 0;
+}
+
+String? _extractVideoId(String reqPath, String suffix) {
+  const String prefix = '/api/library/videos/';
+  final String fullSuffix = '/$suffix';
+  if (!reqPath.startsWith(prefix)) return null;
+  if (!reqPath.endsWith(fullSuffix)) return null;
+  final String id =
+      reqPath.substring(prefix.length, reqPath.length - fullSuffix.length);
+  if (id.isEmpty) return null;
+  // 只拒 `..`（路径穿越），允许 `/`（bookUid 形如 video/xxx）
+  if (id.contains('..') || id.contains('\\')) return null;
+  return id;
+}
+
+/// 从请求 header 读一个 URL-encoded 值并解码（HTTP header 只收 ASCII，非 ASCII 值
+/// 由 client 端 [Uri.encodeComponent] 编码）。缺失/空返回 null；解码失败退回原文。
+String? _decodeHeaderValue(shelf.Request request, String name) {
+  final String? raw = request.headers[name];
+  if (raw == null || raw.isEmpty) return null;
+  try {
+    return Uri.decodeComponent(raw);
+  } catch (_) {
+    return raw;
+  }
+}
+
+File? _coverFile(String? path) {
+  if (path == null || path.isEmpty) return null;
+  final File file = File(path);
+  return file.existsSync() ? file : null;
+}
+
+/// MIME 推断收敛到 hibiki_core 单一映射表 [mimeTypeForFilePath]（命名统一轮 G8）。
+/// 旧本地 switch 副本缺 `.webp` → webp 封面按 application/octet-stream 下发，
+/// 对端 WebView 拒绝内联渲染（BUG-1122）；查共享表后随表修复。保留薄 shim 供
+/// 本文件既有调用方。
+String _guessContentType(String filePath) => mimeTypeForFilePath(filePath);
+
 class FushiSyncServer {
   FushiSyncServer({
     required String syncDataDir,
@@ -213,12 +270,6 @@ class FushiSyncServer {
   final Map<String, _VideoStreamToken> _videoStreamTokens =
       <String, _VideoStreamToken>{};
 
-  /// BUG-1568（BUG-908(a) 的视频同形问题）：视频流 token 的数量上限。签发侧
-  /// （GET /streamurl）先按 TTL prune、再淘汰最旧者收束到上限内。此前 prune 只挂在
-  /// GET /stream 消费侧——只 GET /streamurl 却从不取流的调用者会让
-  /// [_videoStreamTokens] 无界堆积（6 小时 TTL 内每次签发都是净增长）。
-  static const int _maxVideoStreamTokens = 128;
-
   /// BUG-908(d)：WebDAV 写操作（PUT / MKCOL / DELETE）的按路径串行闸门。key 是目标
   /// 文件系统绝对路径，value 是该路径上最近一次写的完成 future。新的写先 await 同一
   /// 路径上前一次写的 future 再执行——同一路径的写串行、不同路径可并行。读操作
@@ -229,14 +280,6 @@ class FushiSyncServer {
   /// 内存态、不落盘；server 重启即清空（半截配对作废，安全侧）。
   final Map<String, FushiPairSession> _pairSessions =
       <String, FushiPairSession>{};
-
-  /// TODO-961 M1（会话 TTL/上限，对照 audio/video token 的 prune 模式）：只发
-  /// pair/v2 却不 confirm 的攻击者会留下永久驻留会话（慢速 DoS）。每个会话
-  /// [_pairSessionTtl] 后过期，超过 [_maxPairSessions] 时先 prune 再淘汰最旧者。
-  /// TTL 与现有 host 60s 自动 deny（[FushiSyncServerController]）对齐：会话生命周期
-  /// 不应长于一次审批窗口（留 90s 余量覆盖审批弹窗 + 用户输 PIN）。
-  static const Duration _pairSessionTtl = Duration(seconds: 90);
-  static const int _maxPairSessions = 64;
 
   /// TODO-961 M3：PIN 爆破限速器。按来源（client 自报 deviceId，回退来源 IP）聚合
   /// PIN 校验失败；同一来源在滑动窗口内累计到阈值即锁定退避一段时间，期间 confirm
@@ -379,155 +422,11 @@ class FushiSyncServer {
     };
   }
 
-  shelf.Middleware _authMiddleware() {
-    return (shelf.Handler innerHandler) {
-      return (shelf.Request request) async {
-        if (request.method == 'OPTIONS') return innerHandler(request);
-        // Pairing is the one unauthenticated route: the client has no token
-        // yet — that is exactly what it is fetching. Gating is done by the
-        // pairing window inside _handlePair, not by Basic auth.
-        if (request.url.path == 'api/pair') return innerHandler(request);
-        // TODO-961 M1: v2 配对（含 confirm）同样无 token——这正是 client 在获取的
-        // 东西。门控由 PIN proof + host 人工确认完成，不靠 Basic auth。
-        if (request.url.path == 'api/pair/v2' ||
-            request.url.path == 'api/pair/v2/confirm') {
-          return innerHandler(request);
-        }
-        // TODO-963 M2: /api/ping 是无鉴权的轻量探测端点——手动输入 IP 的 client 在
-        // 配对前（还没有 token）用它确认地址可达 + 取 host 能力/指纹做 TOFU。只读、
-        // 不泄漏数据，与 /api/pair 同属配对前公开面。
-        if (request.url.path == 'api/ping') return innerHandler(request);
-        // Video stream paths are exempted from Basic auth to allow media_kit
-        // to play via a plain URL. Token validation happens inside the handler.
-        // Only the /stream sub-path is exempted; /streamurl, /subtitle, and the
-        // video list still require Basic auth.
-        if (_isVideoStreamPath(request.url.path)) return innerHandler(request);
-        // Remote lookup audio file URLs are handed to platform audio players,
-        // which issue a bare GET without Authorization. The lookup endpoint
-        // stays authenticated; the file endpoint is guarded by an opaque,
-        // short-lived in-memory id in _handleAudioFile.
-        if (_isLookupAudioFilePath(request.url.path)) {
-          return innerHandler(request);
-        }
-        // TODO-1215: the dictionary media endpoint is fetched by a web page
-        // <img src> GET, which carries no Authorization header (same as
-        // /api/lookup/audio/file). It authenticates via a ?token= query
-        // param instead; a valid token passes here, an invalid/absent one
-        // falls through to the Basic check below (in-app / Basic clients
-        // still work) and ultimately 401s.
-        if (_isDictionaryMediaPath(request.url.path) &&
-            _dictionaryMediaTokenValid(request)) {
-          return innerHandler(request);
-        }
-        final auth = request.headers['authorization'];
-        if (auth == null || !await _validateAuth(auth)) {
-          return shelf.Response(401,
-              headers: {'WWW-Authenticate': 'Basic realm="Fushi Sync"'});
-        }
-        return innerHandler(request);
-      };
-    };
-  }
-
-  /// 判断 [urlPath]（即 request.url.path，不含前导 `/`）是否为视频流路径
-  /// （`api/library/videos/<id>/stream`，id 非空，id 可含 `/`）。
-  static bool _isVideoStreamPath(String urlPath) {
-    const String prefix = 'api/library/videos/';
-    const String suffix = '/stream';
-    if (!urlPath.startsWith(prefix)) return false;
-    if (!urlPath.endsWith(suffix)) return false;
-    final String idPart =
-        urlPath.substring(prefix.length, urlPath.length - suffix.length);
-    return idPart.isNotEmpty;
-  }
-
-  static bool _isLookupAudioFilePath(String urlPath) =>
-      urlPath == 'api/lookup/audio/file';
-
-  Future<bool> _validateAuth(String header) async {
-    final String? password = _basicPassword(header);
-    if (password == null) return false;
-    try {
-      // 兼容路径：共享 [_token] 仍受理（未重新配对的老设备继续可用，Never break
-      // userspace）。常量时间比较防计时侧信道泄漏 token 前缀。
-      if (_constantTimeEquals(
-        Uint8List.fromList(utf8.encode(password)),
-        Uint8List.fromList(utf8.encode(_token)),
-      )) {
-        return true;
-      }
-      // TODO-961 M1b：再比对任一未吊销的 per-peer token。惰性加载 + 缓存，避免每请求
-      // 打库；配对/吊销经 [invalidatePeerTokenCache] 清缓存促重载。
-      final Set<String> peerTokens = await _peerTokens();
-      final Uint8List pw = Uint8List.fromList(utf8.encode(password));
-      for (final String peerToken in peerTokens) {
-        if (_constantTimeEquals(
-            pw, Uint8List.fromList(utf8.encode(peerToken)))) {
-          return true;
-        }
-      }
-      return false;
-    } catch (_) {
-      return false;
-    }
-  }
-
-  /// Sensitive service configuration is stricter than ordinary library APIs:
-  /// it accepts only a currently paired device token, never the legacy shared
-  /// server token.
-  Future<bool> _validatePeerAuth(String? header) async {
-    if (header == null) return false;
-    final String? password = _basicPassword(header);
-    if (password == null) return false;
-    final Uint8List supplied = Uint8List.fromList(utf8.encode(password));
-    for (final String peerToken in await _peerTokens()) {
-      if (_constantTimeEquals(
-        supplied,
-        Uint8List.fromList(utf8.encode(peerToken)),
-      )) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  static String? _basicPassword(String header) {
-    if (!header.startsWith('Basic ')) return null;
-    try {
-      final String decoded =
-          utf8.decode(base64Decode(header.substring('Basic '.length)));
-      final int colonIdx = decoded.indexOf(':');
-      return colonIdx < 0 ? null : decoded.substring(colonIdx + 1);
-    } catch (_) {
-      return null;
-    }
-  }
-
-  /// 当前有效 per-peer token 集合（惰性加载 + 缓存）。无 provider 时为空集——只认
-  /// 共享 [_token]。
-  Future<Set<String>> _peerTokens() async {
-    final Set<String>? cached = _cachedPeerTokens;
-    if (cached != null) return cached;
-    final Future<Set<String>> Function()? provider = pairedPeerTokensProvider;
-    final Set<String> loaded = provider == null ? <String>{} : await provider();
-    _cachedPeerTokens = loaded;
-    return loaded;
-  }
-
   /// TODO-961 M1b：配对新增一台设备 / 吊销一台设备后调用，清 per-peer token 缓存，
   /// 使下一次 [_validateAuth] 从 [pairedPeerTokensProvider] 重新拉取最新集合。
   /// controller 在 upsert / revoke 后调它，保证吊销即时生效、新 token 立即受理。
   void invalidatePeerTokenCache() {
     _cachedPeerTokens = null;
-  }
-
-  static bool _constantTimeEquals(Uint8List a, Uint8List b) {
-    final len = a.length > b.length ? a.length : b.length;
-    var result = a.length ^ b.length;
-    for (var i = 0; i < len; i++) {
-      result |= (i < a.length ? a[i] : 0) ^ (i < b.length ? b[i] : 0);
-    }
-    return result == 0;
   }
 
   Future<shelf.Response> _handleRequest(shelf.Request request) async {
@@ -670,1848 +569,13 @@ class FushiSyncServer {
     }
   }
 
-  /// BUG-908(d)：把 [action] 串到 [fsPath] 的写链尾部——同一路径上的写严格 FIFO 串行，
-  /// 不同路径互不阻塞。实现是「每路径一条链式 future」：新写先取当前链尾 [prev]，把自己
-  /// 的完成 future 挂成新链尾，await [prev] 后再执行 [action]，最后 complete 自己让后继
-  /// 继续。永远只 await 单一路径上的前驱、绝不在持有一把锁时去取另一把，故不会死锁；
-  /// 收尾时若自己仍是链尾就从 map 摘除，避免闲置路径无界堆积。
-  Future<T> _serializeDavWrite<T>(
-      String fsPath, Future<T> Function() action) async {
-    final Future<void> prev = _davWriteChain[fsPath] ?? Future<void>.value();
-    final Completer<void> done = Completer<void>();
-    _davWriteChain[fsPath] = done.future;
-    try {
-      // 等前一次同路径写完成后再动手。前驱失败也不应连累后继，故吞掉其异常。
-      await prev.catchError((Object _) {});
-      return await action();
-    } finally {
-      done.complete();
-      // 若期间没有后继把链尾替换掉，说明该路径已空闲，摘除以防 map 无界增长。
-      if (identical(_davWriteChain[fsPath], done.future)) {
-        _davWriteChain.remove(fsPath);
-      }
-    }
-  }
-
-  Future<shelf.Response> _handlePair(shelf.Request request) async {
-    if (request.method.toUpperCase() != 'POST') return shelf.Response(405);
-    final Future<bool> Function(FushiPairRequest)? approve = onPairRequest;
-    // No UI wired to approve → never hand out the token unattended. A distinct
-    // reason lets the client say "peer not ready" instead of "peer declined".
-    if (approve == null) return _pairDenied('unavailable');
-    // BUG-1555：v1 没有 PIN 概念——它只靠 host 点一下「允许」就发权限最大的**共享**
-    // token（不可逐台吊销）。v2 为公网 / 跨网段入站强制 PIN + HMAC，而 v1 同一条
-    // 入站链路完全绕开这套：攻击者只需改发 /api/pair，host 屏上弹的就是一个普通
-    // 「某设备请求配对」框，误点一次即永久失守。故用与 v2 **同一判据**
-    // ([FushiPairingProtocol.computePinRequired]) 前置拦截：本会话必须 PIN 时直接
-    // 拒 v1（reason='upgrade_required'，让 client 提示升级），压根不弹审批框。
-    //
-    // 兼容性：LAN 内且 host 未开「LAN 也要 PIN」时（默认）v1 行为逐字不变，
-    // 旧客户端继续可配对；变的只是「本就该要 PIN 的会话」——那些会话以前能拿到
-    // 共享 token 本身就是漏洞，现在必须走 v2（Hibiki 自己的 client 早已先试 v2，
-    // 只在对端不支持 v2 时才回落 v1）。
-    final String? pairRemote = _remoteAddress(request);
-    final bool v1PinRequired = FushiPairingProtocol.computePinRequired(
-      isLanPeer: FushiPairingProtocol.isPrivateLanAddress(pairRemote),
-      lanRequiresPin:
-          await (lanRequiresPinProvider?.call() ?? Future<bool>.value(false)),
-    );
-    if (v1PinRequired) return _pairDenied('upgrade_required');
-    String? name;
-    final Map<String, dynamic>? body = await readJsonObjectBody(request);
-    final String? reported = body?['name']?.toString().trim();
-    // Reject a "localhost"/loopback advertisement so it is never stored as this
-    // peer's device name — the paired-devices list would otherwise show
-    // "localhost" instead of a real name (TODO-1356).
-    if (reported != null &&
-        reported.isNotEmpty &&
-        !isMeaninglessDeviceName(reported)) {
-      name = reported;
-    }
-    final bool approved = await approve(FushiPairRequest(
-      deviceName: name,
-      remoteAddress: pairRemote,
-    ));
-    if (!approved) return _pairDenied('declined');
-    return jsonResponse(<String, dynamic>{'token': _token});
-  }
-
-  /// A 403 carrying a machine-readable [reason] ('declined' | 'unavailable' |
-  /// 'upgrade_required' | 'expired') so the client can distinguish a real
-  /// refusal from a peer that has no approval handler wired, from a v1 client
-  /// that this host refuses to pair without a PIN (BUG-1555), and from a
-  /// pairing session that timed out (BUG-1556). Older peers reply with a
-  /// plain-text body instead, which the client treats as 'unavailable'.
-  shelf.Response _pairDenied(String reason) => shelf.Response(
-        403,
-        body: jsonEncode(<String, String>{'reason': reason}),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-
-  /// TODO-961 M1: POST /api/pair/v2 {name, clientNonce} → 200 {sessionId,
-  /// pinRequired, hostNonce}。仅创建会话、决定是否需要 PIN，并把 host 生成的 PIN
-  /// 喂给审批弹窗显示——此阶段 **不** 派 token、**不** 弹「允许/拒绝」（那在
-  /// confirm 阶段，双重确认）。PIN 绝不进响应 body。
-  Future<shelf.Response> _handlePairV2(shelf.Request request) async {
-    if (request.method.toUpperCase() != 'POST') return shelf.Response(405);
-    // No approval UI wired → never start a pairing handshake unattended.
-    if (onPairRequest == null) return _pairDenied('unavailable');
-    final Map<String, dynamic>? body = await readJsonObjectBody(request);
-    final String? clientNonce = body?['clientNonce']?.toString();
-    if (clientNonce == null || clientNonce.trim().isEmpty) {
-      return shelf.Response(400, body: 'Missing clientNonce');
-    }
-    final String? reportedName = body?['name']?.toString().trim();
-    // Drop a "localhost"/loopback advertisement (never a real device name) so it
-    // is not persisted as the peer's name in the paired-devices list (TODO-1356).
-    final String? deviceName = (reportedName != null &&
-            reportedName.isNotEmpty &&
-            !isMeaninglessDeviceName(reportedName))
-        ? reportedName
-        : null;
-    // TODO-961 M1b: client 自报稳定 deviceId（per-peer token 落库的 UNIQUE 身份）。
-    // 旧 client 不带此字段 → null → confirm 回退共享 token（兼容）。
-    final String? reportedDeviceId = body?['clientDeviceId']?.toString().trim();
-    final String? clientDeviceId =
-        (reportedDeviceId != null && reportedDeviceId.isNotEmpty)
-            ? reportedDeviceId
-            : null;
-    final String? remote = _remoteAddress(request);
-
-    final bool isLanPeer = FushiPairingProtocol.isPrivateLanAddress(remote);
-    final bool lanRequiresPin =
-        await (lanRequiresPinProvider?.call() ?? Future<bool>.value(false));
-    final bool pinRequired = FushiPairingProtocol.computePinRequired(
-      isLanPeer: isLanPeer,
-      lanRequiresPin: lanRequiresPin,
-    );
-
-    final String sessionId = FushiPairingProtocol.generateNonce();
-    final String hostNonce = FushiPairingProtocol.generateNonce();
-    final FushiPairSession session = FushiPairSession(
-      sessionId: sessionId,
-      clientNonce: clientNonce,
-      hostNonce: hostNonce,
-      // PIN 先用安全随机兜底，下面交给 host UI 供给器（若接线）覆盖为屏显值。
-      pin: FushiPairingProtocol.generatePin(),
-      pinRequired: pinRequired,
-      deviceName: deviceName,
-      remoteAddress: remote,
-      createdAt: _now(),
-      clientDeviceId: clientDeviceId,
-    );
-    // 先 prune 过期会话 + 守上限：杜绝「只发 pair/v2 不 confirm」的慢速 DoS 把
-    // _pairSessions 撑爆（对照 audio/video token 的 prune 模式）。
-    _prunePairSessions();
-    _enforcePairSessionCap();
-    // 同步回收已冷却的 PIN 失败记录，防限速器内存随开会话数无界增长。
-    _pinRateLimiter.prune(_now());
-
-    // host UI 供给器返回真正显示给用户的 PIN（同值用于 confirm 重算比对）。未接线
-    // 时保留随机兜底 PIN——它不显示给任何人，故 pinRequired 会话无法被 confirm（拒）。
-    final String shownPin = onPairPinGenerated?.call(session) ?? session.pin;
-
-    // TODO-1296 / BUG-592: pinRequired（公网 / 跨网段 / host 要求 PIN）会话在 CREATE
-    // 阶段就弹 host 审批——审批弹窗会显示本会话 PIN，让 client 被要求输入前 host 屏上
-    // 已经有 PIN 可读。修复「公网配对根本看不到 PIN」的时序死锁：旧实现只在 confirm 且
-    // pinProof 校验通过后才弹审批显示 PIN，而 client 必须先输对 PIN 才能过校验 → PIN 永
-    // 远不显示、配对永远走不通。免 PIN 会话（LAN 自动发现且 host 允许免 PIN）审批仍留在
-    // confirm（本就无 PIN 可显示，行为零变化，Never break userspace）。
-    if (pinRequired) {
-      final bool approved = await onPairRequest!(FushiPairRequest(
-        deviceName: deviceName,
-        remoteAddress: remote,
-        // pinVerified 尚未校验（那在 confirm）；pinRequired=true 让审批弹窗显示 PIN。
-        pinVerified: null,
-        pinRequired: true,
-      ));
-      if (!approved) return _pairDenied('declined');
-    }
-
-    // BUG-1556：TTL 从**会话真正开始可用的那一刻**起算，而不是请求入口。
-    // 上面那句 `await onPairRequest!` 是人手审批（host 手机在口袋里 / 用户正在
-    // 忙），完全可能超过整个 90s TTL；若沿用请求入口的时刻，会话会在被写进
-    // [_pairSessions] 的那一刻就已经过期，client 紧接着的 confirm 必被 prune 掉，
-    // 且当时还被报成「对端拒绝」——host 分明刚点了允许。client 真正能开始
-    // 输 PIN 的起点就是审批通过，故 TTL 从此刻起算。
-    final FushiPairSession stored = FushiPairSession(
-      sessionId: sessionId,
-      clientNonce: clientNonce,
-      hostNonce: hostNonce,
-      pin: shownPin,
-      pinRequired: pinRequired,
-      deviceName: deviceName,
-      remoteAddress: remote,
-      createdAt: _now(),
-      clientDeviceId: clientDeviceId,
-    );
-    _pairSessions[sessionId] = stored;
-
-    // 响应只含 sessionId / pinRequired / hostNonce —— 绝不含 PIN 明文。
-    return jsonResponse(<String, dynamic>{
-      'sessionId': sessionId,
-      'pinRequired': pinRequired,
-      'hostNonce': hostNonce,
-    });
-  }
-
-  /// TODO-961 M1: POST /api/pair/v2/confirm {sessionId, pinProof} → 200 {token,
-  /// hostFingerprint}。校验 pinProof（双 nonce HMAC），通过后 **仍** 需 host 人工
-  /// 点允许（双重确认）才派 token。同一 sessionId 二次 confirm（重放）一律拒。
-  Future<shelf.Response> _handlePairConfirm(shelf.Request request) async {
-    if (request.method.toUpperCase() != 'POST') return shelf.Response(405);
-    final Future<bool> Function(FushiPairRequest)? approve = onPairRequest;
-    if (approve == null) return _pairDenied('unavailable');
-    final Map<String, dynamic>? body = await readJsonObjectBody(request);
-    final String? sessionId = body?['sessionId']?.toString();
-    if (sessionId == null || sessionId.trim().isEmpty) {
-      return shelf.Response(400, body: 'Missing sessionId');
-    }
-    // BUG-1556：先**查**再 prune——过期与未知是两回事。旧实现先 prune 再查，
-    // 超时的会话在查之前就没了，于是一律报 'declined'（对端拒绝）——用户
-    // 去查「对方为什么拒绝我」，而真相是「你输 PIN 太慢，会话过期了」，排查
-    // 方向全错。现在过期返回专用 reason='expired'，client 据此说人话。
-    final FushiPairSession? found = _pairSessions[sessionId];
-    if (found != null &&
-        found.createdAt.isBefore(_now().subtract(_pairSessionTtl))) {
-      _pairSessions.remove(sessionId);
-      // 过期会话同样消耗掉它的 PIN：host 屏上那个常驻 PIN 弹窗该收起，
-      // 否则用户盯着一个已经没用的 PIN 继续等（对齐 pinRequired 成功路径）。
-      if (found.pinRequired) onPairSessionResolved?.call();
-      _prunePairSessions();
-      _pinRateLimiter.prune(_now());
-      return _pairDenied('expired');
-    }
-    // 再清掉其余过期会话（不靠后续 create 触发）。
-    _prunePairSessions();
-    // 同步回收已冷却的 PIN 失败记录（锁定中的保留到期满），防限速器内存泄漏。
-    _pinRateLimiter.prune(_now());
-    final FushiPairSession? session = _pairSessions[sessionId];
-    // 未知会话（伪造 / 已被早前的 prune 清走）或已被消费过（重放）→ 拒。
-    // consumed 防同 nonce 二次提交。
-    if (session == null || session.consumed) {
-      return _pairDenied('declined');
-    }
-    // 单次消费：无论本次成功失败，会话即作废，杜绝 nonce 重放。
-    session.consumed = true;
-    _pairSessions.remove(sessionId);
-
-    // TODO-1330 / BUG：pinRequired 会话一旦 confirm 到达，说明 client 已读到 host 屏上
-    // 的 PIN（用它算了 proof），host 那个常驻 PIN 弹窗就该收起——无论本次 proof 对错
-    // （PIN 已一次性消费，重试要走新会话拿新 PIN）。在此单点触发，避开后面多个 return
-    // 分支各自补一遍。免 PIN 会话没有常驻弹窗，不触发。
-    if (session.pinRequired) onPairSessionResolved?.call();
-
-    // TODO-961 M3：本会话来源标识，供爆破限速按来源聚合失败计数。优先 client 自报的
-    // 稳定 deviceId，回退来源 IP；二者都缺时为 null → 无稳定身份可锁，退化为不限速的
-    // 单会话路径（该路径已被 session.consumed 单次消费保护，一个会话只能撞一次）。
-    final String? sourceKey = _pinRateLimitSourceKey(session);
-    if (session.pinRequired) {
-      // 先查锁定态（再触碰 PIN 比对）：锁定则直接拒，杜绝继续撞。锁定判定只看来源与
-      // 时钟，与 PIN 内容完全无关，故不引入「前缀正确就更慢」的计时侧信道。
-      if (sourceKey != null && _pinRateLimiter.isLockedOut(sourceKey, _now())) {
-        return _pairRateLimited();
-      }
-      final String? pinProof = body?['pinProof']?.toString();
-      if (pinProof == null || pinProof.trim().isEmpty) {
-        // 缺 proof 同样计一次失败：否则攻击者可用「开会话→空 proof」零成本探测锁定态
-        // 之外的东西。记后若已锁定，返回 429 让 client 知道被限速。
-        if (sourceKey != null &&
-            _pinRateLimiter.recordFailure(sourceKey, _now())) {
-          return _pairRateLimited();
-        }
-        // 401：PIN 校验未通过（缺 proof）——不再问 host（不弹窗）。
-        return _pairUnauthorized();
-      }
-      final bool ok = FushiPairingProtocol.verifyPinProof(
-        pin: session.pin,
-        clientNonce: session.clientNonce,
-        hostNonce: session.hostNonce,
-        submittedProof: pinProof,
-      );
-      if (!ok) {
-        // 记一次来源级失败；若因此达阈值进入锁定，返回 429（限速），否则 401（PIN 错）。
-        if (sourceKey != null &&
-            _pinRateLimiter.recordFailure(sourceKey, _now())) {
-          return _pairRateLimited();
-        }
-        return _pairUnauthorized();
-      }
-    }
-
-    // TODO-1296 / BUG-592: pinRequired 会话的 host 审批已在 CREATE 阶段完成——会话能
-    // 存在于 _pairSessions 即代表 host 当时已点允许（见 _handlePairV2），故此处不再二次
-    // 弹窗，只凭 pinProof 校验通过即派 token（双重确认 = 早前的人工允许 + 此刻的 proof
-    // 校验，两者仍缺一不可）。免 PIN 会话（pinRequired=false）没有 CREATE 阶段审批，仍在
-    // 此弹审批（无 PIN 可显示，行为不变）。
-    if (!session.pinRequired) {
-      final bool approved = await approve(FushiPairRequest(
-        deviceName: session.deviceName,
-        remoteAddress: session.remoteAddress,
-        pinVerified: true,
-        pinRequired: false,
-      ));
-      if (!approved) return _pairDenied('declined');
-    }
-
-    // TODO-961 M3：成功配对 → 清零该来源的 PIN 失败计数与锁定态（不株连未来尝试）。
-    if (sourceKey != null) _pinRateLimiter.recordSuccess(sourceKey);
-
-    // TODO-961 M1b：per-peer token 派发。仅当 host 接线了落库回调（onPeerPaired）
-    // **且** client 上报了稳定 deviceId 时，才生成本设备专属 token 并写库、回给该
-    // client。任一缺失（纯协议单测无 DB / 旧 client 不上报 deviceId）则回退共享
-    // [_token]——既有行为零变化、老设备继续可配对（Never break userspace）。
-    final String issuedToken = await _issuePeerTokenOrFallback(session);
-    return jsonResponse(<String, dynamic>{
-      'token': issuedToken,
-      if (_securityContext != null && _hostFingerprint != null)
-        'hostFingerprint': _hostFingerprint,
-    });
-  }
-
-  /// confirm 成功后派发访问 token：有 [onPeerPaired] 回调且会话带 clientDeviceId
-  /// 时生成 per-peer token、经回调落库并清 token 缓存（吊销/新增即时生效），返回该
-  /// token；否则回退共享 [_token]（无 DB 接线 / 旧 client 无 deviceId 的兼容路径）。
-  Future<String> _issuePeerTokenOrFallback(FushiPairSession session) async {
-    final Future<void> Function(FushiPairedPeerRegistration)? persist =
-        onPeerPaired;
-    final String? peerId = session.clientDeviceId?.trim();
-    if (persist == null || peerId == null || peerId.isEmpty) {
-      return _token;
-    }
-    final String peerToken = generateToken();
-    await persist(FushiPairedPeerRegistration(
-      peerId: peerId,
-      token: peerToken,
-      deviceName: session.deviceName,
-      remoteAddress: session.remoteAddress,
-    ));
-    // 新 token 立即受理：清缓存促下次 auth 从 provider 重载（含刚写入的这行）。
-    invalidatePeerTokenCache();
-    return peerToken;
-  }
-
-  /// 401，机器可读 reason='pin'：PIN proof 校验未通过。与 403/declined 区分，让
-  /// client 提示「PIN 错误，请重输」而非「对端拒绝」。绝不在 body 里回显任何 PIN。
-  shelf.Response _pairUnauthorized() => shelf.Response(
-        401,
-        body: jsonEncode(<String, String>{'reason': 'pin'}),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-
-  /// TODO-961 M3：429，机器可读 reason='rate_limited'：该来源 PIN 失败过多已被锁定
-  /// 退避。与 401/pin 区分，让 client 提示「尝试过多，请稍后再试」而非「PIN 错误」。
-  /// 绝不回显剩余锁定时长的精确值以外的信息，也绝不泄露 PIN 是否部分正确。
-  shelf.Response _pairRateLimited() => shelf.Response(
-        429,
-        body: jsonEncode(<String, String>{'reason': 'rate_limited'}),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-
-  /// TODO-961 M3：本会话在 PIN 爆破限速里的来源标识。优先 client 自报的稳定
-  /// deviceId（同一物理设备换 IP 也锁得住），回退请求来源 IP。二者都缺（无稳定身份）
-  /// 时返回 null → 调用方退化为不限速的单会话路径（已由 consumed 单次消费保护）。
-  static String? _pinRateLimitSourceKey(FushiPairSession session) {
-    final String? deviceId = session.clientDeviceId?.trim();
-    if (deviceId != null && deviceId.isNotEmpty) return 'dev:$deviceId';
-    final String? remote = session.remoteAddress?.trim();
-    if (remote != null && remote.isNotEmpty) return 'ip:$remote';
-    return null;
-  }
-
-  /// Source IP of the request's TCP connection, or null when shelf_io did not
-  /// attach connection info (e.g. some test harnesses).
-  static String? _remoteAddress(shelf.Request request) {
-    final Object? info = request.context['shelf.io.connection_info'];
-    if (info is HttpConnectionInfo) return info.remoteAddress.address;
-    return null;
-  }
-
-  Future<shelf.Response> _handleLookupApi(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    if (reqPath == '/api/lookup/dictionary') {
-      if (method != 'POST') return shelf.Response(405);
-      return _handleDictionaryLookup(request);
-    }
-    if (reqPath == '/api/lookup/audio') {
-      if (method != 'POST') return shelf.Response(405);
-      return _lookupRoutes.handleAudioLookup(request);
-    }
-    if (reqPath == '/api/lookup/audio/file') {
-      if (method != 'GET' && method != 'HEAD') return shelf.Response(405);
-      return _lookupRoutes.handleAudioFile(request, headOnly: method == 'HEAD');
-    }
-    return shelf.Response.notFound('Not found');
-  }
-
-  Future<shelf.Response> _handleDictionaryLookup(shelf.Request request) async {
-    final FushiRemoteLookupService? service = _remoteLookupService;
-    if (service == null) return shelf.Response.notFound('Remote lookup off');
-    final Map<String, dynamic>? body = await readJsonObjectBody(request);
-    if (body == null) return shelf.Response(400, body: 'Invalid JSON');
-    // 契约与 YomitanApiServer 共享（BUG-530，单一真相源）。
-    return jsonResponse(await buildRemoteDictionaryLookupResponse(
-      body,
-      lookup: service,
-      history: _historyService,
-    ));
-  }
-
-  static bool _isDictionaryMediaPath(String urlPath) =>
-      urlPath == 'api/media/dictionary';
-
-  /// Validates that [request]'s query token (?token= etc.) equals [_token].
-  /// Used to admit the dictionary media endpoint without Basic auth (a web
-  /// page <img> GET has no Authorization header). Constant-time comparison.
-  bool _dictionaryMediaTokenValid(shelf.Request request) {
-    for (final String name in _kDictionaryMediaTokenParams) {
-      final String? value = request.url.queryParameters[name];
-      if (value != null &&
-          _constantTimeEquals(
-            Uint8List.fromList(utf8.encode(value)),
-            Uint8List.fromList(utf8.encode(_token)),
-          )) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /// TODO-1215: GET /api/media/dictionary?dictionary=<name>&path=<rel>&token=<t>
-  /// -- dictionary gaiji / pitch-accent SVG (etc.) media bytes. The browser
-  /// extension popup rewrites a term's <img src> from the in-app image:// URL
-  /// to this endpoint (a real browser has no image:// handler). Bytes are
-  /// resolved by the injected [_dictionaryMediaProvider] (bridged to
-  /// FushiDicts.getMediaFile in-app); the MIME reuses the same
-  /// [dictionaryMediaMimeType] as the app scheme handler.
-  shelf.Response _handleDictionaryMedia(shelf.Request request, bool headOnly) {
-    final Uint8List? Function(String, String)? provider =
-        _dictionaryMediaProvider;
-    if (provider == null) return shelf.Response.notFound('Media off');
-    final String dictionary = request.url.queryParameters['dictionary'] ?? '';
-    final String path = normalizeDictionaryMediaPath(
-      request.url.queryParameters['path'] ?? '',
-    );
-    if (dictionary.isEmpty || path.isEmpty) {
-      return shelf.Response.notFound('Not found');
-    }
-    final Uint8List? bytes = provider(dictionary, path);
-    if (bytes == null || bytes.isEmpty) {
-      return shelf.Response.notFound('Not found');
-    }
-    final String mime = dictionaryMediaMimeType(path);
-    return shelf.Response.ok(
-      headOnly ? null : bytes,
-      headers: <String, String>{
-        'Content-Type': mime,
-        'Content-Length': '${bytes.length}',
-        // The extension loads the image cross-origin on a real web page (and
-        // may drawImage it onto a canvas for monochrome recolor); allow CORS
-        // so any read-back path never taints the canvas. Media is
-        // non-sensitive and already token-authenticated.
-        'Access-Control-Allow-Origin': '*',
-      },
-    );
-  }
-
-  // /api/mine、/api/mine/forward、/api/anki/note-type/*、/api/duplicate、
-  // /api/lookup/audio[/file] 的 handler 正文在 [RemoteLookupRoutes]（与
-  // YomitanApiServer 共用）；本文件只留互联独有的端点。
-
-  /// 互联媒体存储优化：客户端（手机）经互联对**主机端** collection.media 做字节级
-  /// 去重——卡片落在主机的 Anki 上，重复媒体也堆在主机，客户端本机根本没有那个
-  /// 目录。未注入挖词 service → 404（旧版主机对新客户端同样 404 → 客户端按
-  /// 「不支持」隐藏区块）；dryRun 类型错 → 400。
-  Future<shelf.Response> _handleAnkiMediaDedup(
-    shelf.Request request,
-    String path,
-  ) async {
-    final FushiRemoteMiningService? svc = _miningService;
-    if (svc == null) return shelf.Response.notFound('Mining off');
-    final Map<String, dynamic>? body = await readJsonObjectBody(request);
-    if (body == null) return shelf.Response(400, body: 'Invalid JSON');
-    try {
-      switch (path) {
-        case '/api/anki/media/dedup/probe':
-          return jsonResponse(
-              await buildAnkiMediaDedupProbeResponse(mining: svc));
-        case '/api/anki/media/dedup/run':
-          return jsonResponse(
-              await buildAnkiMediaDedupRunResponse(body, mining: svc));
-        default:
-          return shelf.Response.notFound('Unknown endpoint');
-      }
-    } on FormatException catch (e) {
-      return shelf.Response(400, body: e.message);
-    }
-  }
-
-  /// TODO-963 M2: 无鉴权轻量探测。手动输入 IP 的 client 在「填 IP → 探测 → 配对」流程
-  /// 里用它确认地址可达、读 host 展示名 + 是否支持 v2 配对 + （TLS 开时）host 证书指纹
-  /// 供 TOFU 钉扎。只读、不含任何数据/凭据。绝不回传 token。
-  shelf.Response _handlePing() {
-    return jsonResponse(<String, dynamic>{
-      // 互联 wire 服务字段：与 client 侧 probeFushiPing 的 app == 'fushi'
-      // 同版本对切（R11 已接受跨版本配对探测互不识别）。
-      'app': 'fushi',
-      'pairing': <String, dynamic>{'v2': true},
-      'tls': <String, dynamic>{
-        'enabled': _securityContext != null,
-        if (_hostFingerprint != null) 'fingerprint': _hostFingerprint,
-      },
-      if (_deviceName != null && _deviceName.isNotEmpty)
-        'deviceName': _deviceName,
-    });
-  }
-
-  Future<shelf.Response> _handleCapabilities() async {
-    final bool lib = _libraryService != null;
-    // 漫画 P3 能力协商：仅接线了 OCR 任务管理器的 host 带 `mangaOcr` 字段；老
-    // host 响应里没有该字段 → client 隐藏「已配对主机」OCR 选项（零破坏）。
-    final Map<String, Object?>? mangaOcr = await _mangaOcrJobs?.capability();
-    return jsonResponse(<String, dynamic>{
-      if (mangaOcr != null) 'mangaOcr': mangaOcr,
-      'liveLibrary': <String, dynamic>{
-        'dictionaries': lib,
-        'books': lib,
-        'audio': lib,
-        'videos': lib,
-        'serviceConfig': _securityContext != null &&
-            _libraryService is InterconnectServiceConfigHost,
-        // 互联「配置文件」（Profile）双向搬运：与 serviceConfig 同门槛（必须 TLS）。
-        // 能力位只说「这台 host 懂这个端点」，不代表此刻允许——端点还会再查一次
-        // 用户开关，关着时返回 403，client 如实报错而不是当成不支持。
-        'profileTransfer': _securityContext != null &&
-            _libraryService is InterconnectProfileHost,
-      },
-      // TODO-961 M1 能力协商（设计稿 §1.1 / §2.5）：老 client 读不到也不崩。
-      'tls': <String, dynamic>{
-        'enabled': _securityContext != null,
-        if (_hostFingerprint != null) 'fingerprint': _hostFingerprint,
-      },
-      'pairing': <String, dynamic>{'v2': true},
-    });
-  }
-
-  /// HBK-AUDIT-012 路径穿越闸门：资产名绝不能含路径分隔符或 `..`，否则能逃出 host
-  /// 的资产根目录（DELETE 最危险）。合法返回 null；非法直接返回要回给客户端的响应。
-  ///
-  /// 此前这段判断在四个域 + 三个 position/progress 子路由里逐字重复了 7 遍。安全闸门
-  /// 靠复制粘贴维持，抄漏一处就是真漏洞——收敛成一处后新端点只能显式调用它。
-  ///
-  /// 注意：视频域**不用**本闸门——视频 id 形如 `video/xxx`，合法地含 `/`，它有自己的
-  /// `_extractVideoId` 校验。不要把视频接进来。
-  shelf.Response? _rejectUnsafeAssetId(String id, String label) {
-    if (id.isEmpty) return shelf.Response.notFound('Missing $label');
-    if (id.contains('/') || id.contains('\\') || id.contains('..')) {
-      return shelf.Response.forbidden('Invalid $label');
-    }
-    return null;
-  }
-
-  /// 「按名字取 / 存 / 删一个资产包」端点的共同骨架：词典 / 书 / 本地音频 / 有声书
-  /// 四个域在这一层**逐字相同**，只差叫什么名字、临时文件用什么扩展名、调 service
-  /// 的哪三个方法。
-  ///
-  /// 此前是四份约 60 行的复制粘贴（TODO-2120），加一个新媒体域就要再抄一遍——而这
-  /// 段代码里含导出缓存 + ETag/Range 续传、上传临时目录的必清理、IOSink 的双重关闭
-  /// 保护，抄漏任何一处都是真事故（泄漏临时目录 / 续传验证器失效 / socket 不回收）。
-  ///
-  /// [id] 必须已经过 [_rejectUnsafeAssetId]。
-  Future<shelf.Response> _serveAssetPackage(
-    shelf.Request request,
-    String method, {
-    required String id,
-    required String cacheKind,
-    required String notFoundMessage,
-    required String tempPrefix,
-    required String tempExtension,
-    required Future<File> Function() export,
-    required Future<void> Function(File tmp) import,
-    required Future<void> Function() delete,
-  }) async {
-    switch (method) {
-      case 'GET':
-        // 经导出缓存 + Range/If-Range：TTL 内的续传钉在同一份字节上（ETag 作验证器）；
-        // 旧 client 不发 Range 收到 200 全量，行为不变。扩展名保留 → Content-Type
-        // 仍由 _guessContentType 按扩展名判定。
-        File file;
-        try {
-          file = await _exportCache.obtain(cacheKind, id, export);
-        } on StateError {
-          return shelf.Response.notFound(notFoundMessage);
-        }
-        return serveFileWithRange(file, request,
-            etag: ExportPackageCache.etagFor(file));
-
-      case 'PUT':
-        final Directory tmpDir =
-            Directory.systemTemp.createTempSync(tempPrefix);
-        final File tmp = File(p.join(tmpDir.path, '$id$tempExtension'));
-        final IOSink sink = tmp.openWrite();
-        try {
-          await request.read().forEach(sink.add);
-          await sink.close();
-          await import(tmp);
-          return shelf.Response(200);
-        } catch (e) {
-          try {
-            await sink.close();
-          } catch (_) {
-            // best-effort
-          }
-          return shelf.Response(500, body: 'Import failed: $e');
-        } finally {
-          try {
-            tmpDir.deleteSync(recursive: true);
-          } catch (_) {
-            // best-effort
-          }
-        }
-
-      case 'DELETE':
-        await delete();
-        return shelf.Response(204);
-
-      default:
-        return shelf.Response(405);
-    }
-  }
-
-  Future<shelf.Response> _handleLibraryDictionaries(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    if (reqPath == '/api/library/dictionaries') {
-      if (method != 'GET') return shelf.Response(405);
-      final List<RemoteDictionaryInfo> list = await svc.listDictionaries();
-      return shelf.Response.ok(
-        jsonEncode(<Map<String, Object?>>[
-          for (final RemoteDictionaryInfo d in list) d.toJson()
-        ]),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-    }
-
-    // reqPath 已在 _handleRequest 经 Uri.decodeFull 解码，此处无需再解码。
-    // 原先的 Uri.decodeComponent 调用会对已解码的 CJK 字符再次解析，
-    // 导致 "Illegal percent encoding in URI"（Dart 不接受非 ASCII 作为
-    // decodeComponent 输入）。直接 substring 即可得到正确的词典名。
-    final String name = reqPath.substring('/api/library/dictionaries/'.length);
-    // HBK-AUDIT-012 路径穿越闸门（收敛到 _rejectUnsafeAssetId），覆盖下面三个方法。
-    final shelf.Response? unsafe =
-        _rejectUnsafeAssetId(name, 'dictionary name');
-    if (unsafe != null) return unsafe;
-
-    return _serveAssetPackage(
-      request,
-      method,
-      id: name,
-      cacheKind: 'dict',
-      notFoundMessage: 'Dictionary not found',
-      tempPrefix: 'fushi_dict_in',
-      tempExtension: '.fushidict',
-      export: () => svc.exportDictionary(name),
-      import: svc.importDictionary,
-      delete: () => svc.deleteDictionary(name),
-    );
-  }
-
-  Future<shelf.Response> _handleLibraryBooks(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    if (reqPath == '/api/library/books') {
-      if (method != 'GET') return shelf.Response(405);
-      final List<RemoteBookInfo> list = await svc.listBooks();
-      return shelf.Response.ok(
-        jsonEncode(<Map<String, Object?>>[
-          for (final RemoteBookInfo b in list)
-            _remoteBookJsonForRequest(b, request)
-        ]),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-    }
-
-    // reqPath 已在 _handleRequest 经 Uri.decodeFull 解码，此处无需再解码。
-    const String bookPrefix = '/api/library/books/';
-    const String coverSuffix = '/cover';
-    if (reqPath.startsWith(bookPrefix) && reqPath.endsWith(coverSuffix)) {
-      if (method != 'GET') return shelf.Response(405);
-      final String coverBookId = reqPath.substring(
-          bookPrefix.length, reqPath.length - coverSuffix.length);
-      final shelf.Response? unsafeCoverBookId =
-          _rejectUnsafeAssetId(coverBookId, 'book title');
-      if (unsafeCoverBookId != null) return unsafeCoverBookId;
-      final File? cover = await _resolveBookCover(svc, coverBookId);
-      if (cover == null) return shelf.Response.notFound('Book cover not found');
-      return serveFileWithRange(cover, request);
-    }
-
-    // GET/PUT /api/library/books/<bookKey>/progress — 跨设备阅读进度（TODO-767）。
-    // GET 让 client 拉取 host 真相源进度；PUT 让 client 上报本端进度（host 取较新者）。
-    // 与 video /position 分支对称，但落 host 自己的 reader_positions DB（非 prefs）。
-    const String progressSuffix = '/progress';
-    if (reqPath.startsWith(bookPrefix) && reqPath.endsWith(progressSuffix)) {
-      final String progressBookKey = reqPath.substring(
-          bookPrefix.length, reqPath.length - progressSuffix.length);
-      final shelf.Response? unsafeProgressBookKey =
-          _rejectUnsafeAssetId(progressBookKey, 'book key');
-      if (unsafeProgressBookKey != null) return unsafeProgressBookKey;
-      switch (method) {
-        case 'GET':
-          final RemoteBookProgress progress =
-              await svc.getBookProgress(progressBookKey);
-          return shelf.Response.ok(
-            jsonEncode(progress.toJson()),
-            headers: <String, String>{'Content-Type': 'application/json'},
-          );
-        case 'PUT':
-          final String body = await request.readAsString();
-          Map<String, dynamic> json;
-          try {
-            json = jsonDecode(body) as Map<String, dynamic>;
-          } catch (_) {
-            return shelf.Response(400, body: 'Invalid JSON');
-          }
-          await svc.putBookProgress(
-            progressBookKey,
-            RemoteBookProgress.fromJson(json.cast<String, Object?>()),
-          );
-          return shelf.Response(200);
-        default:
-          return shelf.Response(405);
-      }
-    }
-
-    final String bookId = reqPath.substring(bookPrefix.length);
-    final shelf.Response? unsafe = _rejectUnsafeAssetId(bookId, 'book title');
-    if (unsafe != null) return unsafe;
-
-    // 注：`.epub` 扩展名保留 → Content-Type 仍是 application/epub+zip
-    // （见 _guessContentType）。
-    return _serveAssetPackage(
-      request,
-      method,
-      id: bookId,
-      cacheKind: 'book',
-      notFoundMessage: 'Book not found',
-      tempPrefix: 'hibiki_book_in',
-      tempExtension: '.epub',
-      export: () => svc.exportBook(bookId),
-      // BUG-1503：body 是裸 .epub，推送方用户改的书名走 header 随行（视频推送的
-      // `X-Hibiki-Video-Title` 同一先例）。旧 client 不发 → 两参 null/0 → 与
-      // 原来的 `svc.importBook` tear-off 逐字同行为，故不必动
-      // [_serveAssetPackage] 的四域共用签名。
-      import: (File epubFile) => svc.importBook(
-        epubFile,
-        displayTitle: _decodeHeaderValue(request, kBookDisplayTitleHeader),
-        displayTitleAt: int.tryParse(
-              request.headers[kBookDisplayTitleAtHeader] ?? '',
-            ) ??
-            0,
-      ),
-      delete: () => svc.deleteBook(bookId),
-    );
-  }
-
-  Map<String, Object?> _remoteBookJsonForRequest(
-    RemoteBookInfo book,
-    shelf.Request request,
-  ) {
-    final Map<String, Object?> json = book.toJson()
-      ..remove('coverUrl')
-      ..remove('hasCover');
-    if (_coverFile(book.coverPath) != null) {
-      json['hasCover'] = true;
-      json['coverUrl'] = request.requestedUri.replace(
-        pathSegments: <String>[
-          'api',
-          'library',
-          'books',
-          book.downloadId,
-          'cover',
-        ],
-        queryParameters: <String, String>{},
-      ).toString();
-    }
-    return json;
-  }
-
-  Future<File?> _resolveBookCover(
-    FushiLibraryHostService service,
-    String bookId,
-  ) async {
-    // 单行直查（bookCoverPath 按 bookKey 优先、title 兜底），不再每张封面重跑
-    // 整份 listBooks()（逐书标签/有声书/合集重活），与 [_resolveVideoCover] 对称。
-    return _coverFile(await service.bookCoverPath(bookId));
-  }
-
-  Future<shelf.Response> _handleLibraryLocalAudio(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    if (reqPath == '/api/library/localaudio') {
-      if (method != 'GET') return shelf.Response(405);
-      final List<RemoteLocalAudioInfo> list = await svc.listLocalAudio();
-      return shelf.Response.ok(
-        jsonEncode(<Map<String, Object?>>[
-          for (final RemoteLocalAudioInfo a in list) a.toJson()
-        ]),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-    }
-
-    // reqPath 已在 _handleRequest 经 Uri.decodeFull 解码，此处无需再解码。
-    final String displayName =
-        reqPath.substring('/api/library/localaudio/'.length);
-    final shelf.Response? unsafe =
-        _rejectUnsafeAssetId(displayName, 'displayName');
-    if (unsafe != null) return unsafe;
-
-    return _serveAssetPackage(
-      request,
-      method,
-      id: displayName,
-      cacheKind: 'localaudio',
-      notFoundMessage: 'Local audio not found',
-      tempPrefix: 'hibiki_localaudio_in',
-      tempExtension: '.localaudio',
-      export: () => svc.exportLocalAudio(displayName),
-      import: svc.importLocalAudio,
-      delete: () => svc.deleteLocalAudio(displayName),
-    );
-  }
-
-  Future<shelf.Response> _handleLibraryAudiobooks(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    if (reqPath == '/api/library/audiobooks') {
-      if (method != 'GET') return shelf.Response(405);
-      final List<RemoteAudiobookInfo> list = await svc.listAudiobooks();
-      return shelf.Response.ok(
-        jsonEncode(<Map<String, Object?>>[
-          for (final RemoteAudiobookInfo ab in list) ab.toJson()
-        ]),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-    }
-
-    // reqPath 已在 _handleRequest 经 Uri.decodeFull 解码，此处无需再解码。
-    // GET/PUT /api/library/audiobooks/<bookKey>/position — 跨设备有声书播放断点
-    // (BUG-471)。GET 让 client 拉取 host 真相源进度；PUT 让 client 上报本端进度
-    // (host 取较新者)。与 video /position 分支对称。必须在下面的整包 bookKey 提取
-    // 之前匹配，否则 `<key>/position` 会被当成含 `/` 的非法 bookKey 拒绝。
-    const String audiobookPrefix = '/api/library/audiobooks/';
-    const String positionSuffix = '/position';
-    if (reqPath.startsWith(audiobookPrefix) &&
-        reqPath.endsWith(positionSuffix)) {
-      final String positionBookKey = reqPath.substring(
-          audiobookPrefix.length, reqPath.length - positionSuffix.length);
-      final shelf.Response? unsafePositionBookKey =
-          _rejectUnsafeAssetId(positionBookKey, 'bookKey');
-      if (unsafePositionBookKey != null) return unsafePositionBookKey;
-      // 先确认该有声书在 host DB 真实存在，防任意 key 写脏 prefs；与视频 position
-      // 先 resolveVideoFile 同语义。BUG-471a：改用廉价的 audiobookExists（单次 DB
-      // 查询）替代旧的 exportAudiobook 打包探测——旧实现每次 GET/PUT position 都把整
-      // 本有声书音频/字幕/封面打成 .fushiaudio 临时文件再删，live sweep 对每本共享
-      // 有声书每轮触发一次造成大量无谓 zip I/O + CPU。
-      try {
-        if (!await svc.audiobookExists(positionBookKey)) {
-          return shelf.Response.notFound('Audiobook not found');
-        }
-      } on ArgumentError {
-        return shelf.Response.forbidden('Invalid bookKey');
-      }
-      switch (method) {
-        case 'GET':
-          final ({int positionMs, int updatedAtMs}) p =
-              await svc.getAudiobookPosition(positionBookKey);
-          return shelf.Response.ok(
-            jsonEncode(<String, Object?>{
-              'positionMs': p.positionMs,
-              'positionUpdatedAtMs': p.updatedAtMs,
-            }),
-            headers: <String, String>{'Content-Type': 'application/json'},
-          );
-        case 'PUT':
-          final String body = await request.readAsString();
-          Map<String, dynamic> json;
-          try {
-            json = jsonDecode(body) as Map<String, dynamic>;
-          } catch (_) {
-            return shelf.Response(400, body: 'Invalid JSON');
-          }
-          final int posMs = (json['positionMs'] as num?)?.toInt() ?? 0;
-          final int updatedAtMs =
-              (json['positionUpdatedAtMs'] as num?)?.toInt() ?? 0;
-          await svc.putAudiobookPosition(positionBookKey, posMs, updatedAtMs);
-          return shelf.Response(200);
-        default:
-          return shelf.Response(405);
-      }
-    }
-
-    // GET/PUT /api/library/audiobooks/<identity>/delay — 有声书调轴跨设备同步
-    // （互联完整支持批次；与视频 /delay、上面的 /position 分支对称）。同样必须在
-    // 整包 bookKey 提取之前匹配。老 host 无此分支 → 404，client best-effort 降级。
-    const String delaySuffix = '/delay';
-    if (reqPath.startsWith(audiobookPrefix) && reqPath.endsWith(delaySuffix)) {
-      final String delayIdentity = reqPath.substring(
-          audiobookPrefix.length, reqPath.length - delaySuffix.length);
-      final shelf.Response? unsafeDelayIdentity =
-          _rejectUnsafeAssetId(delayIdentity, 'bookKey');
-      if (unsafeDelayIdentity != null) return unsafeDelayIdentity;
-      if (svc is! AudiobookDelayHost) {
-        return shelf.Response.notFound('Audiobook delay not supported');
-      }
-      final AudiobookDelayHost delayHost = svc as AudiobookDelayHost;
-      try {
-        if (!await svc.audiobookExists(delayIdentity)) {
-          return shelf.Response.notFound('Audiobook not found');
-        }
-      } on ArgumentError {
-        return shelf.Response.forbidden('Invalid bookKey');
-      }
-      switch (method) {
-        case 'GET':
-          final ({int delayMs, int updatedAtMs}) d =
-              await delayHost.getAudiobookDelay(delayIdentity);
-          return jsonResponse(<String, dynamic>{
-            'delayMs': d.delayMs,
-            'delayUpdatedAtMs': d.updatedAtMs,
-          });
-        case 'PUT':
-          final String body = await request.readAsString();
-          Map<String, dynamic> json;
-          try {
-            json = jsonDecode(body) as Map<String, dynamic>;
-          } catch (_) {
-            return shelf.Response(400, body: 'Invalid JSON');
-          }
-          final int delayMs = (json['delayMs'] as num?)?.toInt() ?? 0;
-          final int updatedAtMs =
-              (json['delayUpdatedAtMs'] as num?)?.toInt() ?? 0;
-          await delayHost.putAudiobookDelay(
-              delayIdentity, delayMs, updatedAtMs);
-          return shelf.Response(200);
-        default:
-          return shelf.Response(405);
-      }
-    }
-
-    final String bookKey = reqPath.substring('/api/library/audiobooks/'.length);
-    final shelf.Response? unsafe = _rejectUnsafeAssetId(bookKey, 'bookKey');
-    if (unsafe != null) return unsafe;
-
-    // 有声书包是 Range 续传的最大受益者（包最大）。导入要带 bookKeyOverride：
-    // 落地时必须钉在 URL 上的这个 key，不能让包里的自述 key 改写身份（BUG-414）。
-    return _serveAssetPackage(
-      request,
-      method,
-      id: bookKey,
-      cacheKind: 'audiobook',
-      notFoundMessage: 'Audiobook not found',
-      tempPrefix: 'hibiki_audiobook_in',
-      tempExtension: '.audiobook',
-      export: () => svc.exportAudiobook(bookKey),
-      import: (File tmp) => svc.importAudiobook(tmp, bookKeyOverride: bookKey),
-      delete: () => svc.deleteAudiobook(bookKey),
-    );
-  }
-
-  // ── 视频端点（P4-2）──────────────────────────────────────────────────────────
-
-  /// 从视频子路径提取视频 id。
-  ///
-  /// [reqPath] 已经过 Uri.decodeFull 解码（含前导 `/`）。
-  /// 格式为 `/api/library/videos/<id>/<suffix>`，其中：
-  /// - [suffix] 为 `stream`、`streamurl`、`subtitle` 或 `cover`
-  /// - id 允许包含 `/`（如 `video/my_film`），但不允许 `..`（路径穿越）
-  ///
-  /// 解析失败（id 为空或含 `..`）时返回 null。
-  /// 解析 `?episode=N` query 成集下标（TODO-885）；缺省 / 非法 / 负数都回退 0
-  /// （= 当前集 / 单视频，向后兼容）。
-  static int _episodeIndexFromRequest(shelf.Request request) {
-    final String? raw = request.url.queryParameters['episode'];
-    if (raw == null) return 0;
-    final int? n = int.tryParse(raw);
-    if (n == null || n < 0) return 0;
-    return n;
-  }
-
-  static String? _extractVideoId(String reqPath, String suffix) {
-    const String prefix = '/api/library/videos/';
-    final String fullSuffix = '/$suffix';
-    if (!reqPath.startsWith(prefix)) return null;
-    if (!reqPath.endsWith(fullSuffix)) return null;
-    final String id =
-        reqPath.substring(prefix.length, reqPath.length - fullSuffix.length);
-    if (id.isEmpty) return null;
-    // 只拒 `..`（路径穿越），允许 `/`（bookUid 形如 video/xxx）
-    if (id.contains('..') || id.contains('\\')) return null;
-    return id;
-  }
-
-  /// 裸视频 id（`/api/library/videos/<id>`，**无** suffix）的提取 + 穿越校验。
-  ///
-  /// PUT（client→host 上传）与 DELETE（client→host 删除）共用这一处：两者都只接
-  /// 「裸 id」，带 suffix 的子路由（cover / streamurl / stream / subtitle / position /
-  /// clipaudio）在上方已被 [_extractVideoId] 消化掉。
-  ///
-  /// 收敛成函数而不是在每个端点里手抄穿越判断——安全闸门靠复制粘贴维持，抄漏一处
-  /// 就是真漏洞（同 [_rejectUnsafeAssetId] 的教训，守卫见
-  /// `test/sync/fushi_sync_server_asset_gate_test.dart`，它按纯文本计数穿越判断的
-  /// 出现次数，所以正文注释里也不要写那个字面量）。视频域不能用那道资产闸门是因为
-  /// 它禁 `/`，而视频 bookUid 合法含 `/`。
-  static String? _extractBareVideoId(String reqPath) {
-    const String prefix = '/api/library/videos/';
-    if (!reqPath.startsWith(prefix)) return null;
-    final String id = reqPath.substring(prefix.length);
-    if (id.isEmpty) return null;
-    if (id.contains('..') || id.contains('\\')) return null;
-    return id;
-  }
-
-  /// GET /api/library/activity — host 最近活动事件（新首页 Activity 面板的互联
-  /// 数据源；display-only，client 不落库）。limit 参数钳制 1..500。老 client 不知
-  /// 道此端点、老 host 对此路径 404（client 侧优雅降级为空列表）。
-  Future<shelf.Response> _handleLibraryActivity(
-    shelf.Request request,
-    String method,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-    if (method != 'GET') return shelf.Response(405);
-    final int limit =
-        (int.tryParse(request.url.queryParameters['limit'] ?? '') ?? 100)
-            .clamp(1, 500);
-    final List<RemoteActivityEvent> events =
-        await svc.listActivityEvents(limit: limit);
-    return shelf.Response.ok(
-      jsonEncode(<Map<String, Object?>>[
-        for (final RemoteActivityEvent e in events) e.toJson(),
-      ]),
-      headers: <String, String>{'Content-Type': 'application/json'},
-    );
-  }
-
-  Future<shelf.Response> _handleLibraryVideos(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    // GET /api/library/videos — 列表（需 Basic 鉴权，中间件已处理）
-    if (reqPath == '/api/library/videos') {
-      if (method != 'GET') return shelf.Response(405);
-      final List<RemoteVideoInfo> list = await svc.listVideos();
-      return shelf.Response.ok(
-        jsonEncode(<Map<String, Object?>>[
-          for (final RemoteVideoInfo v in list)
-            _remoteVideoJsonForRequest(v, request)
-        ]),
-        headers: <String, String>{'Content-Type': 'application/json'},
-      );
-    }
-
-    // GET /api/library/videos/<id>/cover — 视频封面（需 Basic 鉴权）
-    final String? coverId = _extractVideoId(reqPath, 'cover');
-    if (coverId != null) {
-      if (method != 'GET') return shelf.Response(405);
-      final File? cover = await _resolveVideoCover(svc, coverId);
-      if (cover == null) {
-        return shelf.Response.notFound('Video cover not found');
-      }
-      return serveFileWithRange(cover, request);
-    }
-
-    // GET /api/library/videos/<id>/streamurl — 签发短时 token（需 Basic 鉴权）
-    final String? streamUrlId = _extractVideoId(reqPath, 'streamurl');
-    if (streamUrlId != null) {
-      if (method != 'GET') return shelf.Response(405);
-      // TODO-885: 远端播放列表按集——?episode=N 决定流式哪一集（DB-only 反查）。
-      final int episodeIndex = _episodeIndexFromRequest(request);
-      final File? file =
-          await svc.resolveVideoFile(streamUrlId, episodeIndex: episodeIndex);
-      if (file == null) return shelf.Response.notFound('Video not found');
-      // BUG-1568：签发前先按 TTL 清过期，再把数量收束到 [_maxVideoStreamTokens] 内
-      // （淘汰最旧者）。对照 audio token 的 BUG-908(a) 修法：消费侧（GET /stream）的
-      // prune 等不到「只签发不取流」的调用者，上限必须在签发侧强制。
-      _pruneVideoTokens();
-      _enforceVideoTokenCap();
-      final String tokenValue = _generateVideoToken();
-      _videoStreamTokens[tokenValue] = _VideoStreamToken(
-        videoId: streamUrlId,
-        createdAt: _now(),
-        episodeIndex: episodeIndex,
-      );
-      final String encodedId = Uri.encodeFull(streamUrlId);
-      // stream / subtitle URL 都带 episode=N，让 client 取流 / 下字幕命中同一集。
-      final Map<String, String> streamQuery = <String, String>{
-        'token': tokenValue,
-        if (episodeIndex > 0) 'episode': '$episodeIndex',
-      };
-      final Uri streamUri = request.requestedUri.replace(
-        path: '/api/library/videos/$encodedId/stream',
-        queryParameters: streamQuery,
-      );
-      // subtitle URL 不含 token（走 Basic 鉴权），但带 episode=N。
-      final File? sub = await svc.resolveVideoSubtitle(streamUrlId,
-          episodeIndex: episodeIndex);
-      final Uri? subtitleUri = sub != null
-          ? request.requestedUri.replace(
-              path: '/api/library/videos/$encodedId/subtitle',
-              queryParameters: <String, String>{
-                if (episodeIndex > 0) 'episode': '$episodeIndex',
-              },
-            )
-          : null;
-      final List<RemoteVideoEmbeddedSubtitleTrack> embeddedTracks =
-          await _embeddedSubtitleTracksForRequest(
-        file,
-        request,
-        streamUrlId,
-        episodeIndex,
-      );
-      return jsonResponse(<String, dynamic>{
-        'url': streamUri.toString(),
-        'subtitleUrl': subtitleUri?.toString(),
-        if (sub != null) 'subtitleFileName': p.basename(sub.path),
-        if (embeddedTracks.isNotEmpty)
-          'embeddedSubtitleTracks': <Map<String, Object?>>[
-            for (final RemoteVideoEmbeddedSubtitleTrack track in embeddedTracks)
-              track.toJson(),
-          ],
-      });
-    }
-
-    // GET /api/library/videos/<id>/stream — 流式传输（豁免 Basic，靠 token 鉴权）
-    final String? streamId = _extractVideoId(reqPath, 'stream');
-    if (streamId != null) {
-      if (method != 'GET') return shelf.Response(405);
-      _pruneVideoTokens();
-      final String? tokenValue = request.url.queryParameters['token'];
-      if (tokenValue == null || tokenValue.isEmpty) {
-        return shelf.Response(401,
-            body: 'Missing token',
-            headers: <String, String>{'Content-Type': 'text/plain'});
-      }
-      final _VideoStreamToken? tok = _videoStreamTokens[tokenValue];
-      if (tok == null || tok.videoId != streamId) {
-        return shelf.Response(403,
-            body: 'Invalid or expired token',
-            headers: <String, String>{'Content-Type': 'text/plain'});
-      }
-      // TODO-885: 用 token 绑定的集下标反查（token 是 streamurl 签发时定的，client 不能
-      // 自己改集——?episode 只决定 streamurl 阶段，stream 阶段以 token 为准）。
-      final File? file =
-          await svc.resolveVideoFile(streamId, episodeIndex: tok.episodeIndex);
-      if (file == null) return shelf.Response.notFound('Video not found');
-      return serveFileWithRange(file, request);
-    }
-
-    // GET /api/library/videos/<id>/subtitle — 字幕（需 Basic 鉴权，中间件已处理）
-    // PUT 同路径 — client→host 上传该视频的外挂字幕 sidecar（BUG-964，随
-    // syncVideoFiles live push）。后缀（`.srt` / `.ja.srt` …）经
-    // X-Hibiki-Subtitle-Suffix header 上报，服务端白名单校验；老 host 无此分支
-    // 对 PUT 回 405，client 据此优雅降级。
-    final String? subtitleId = _extractVideoId(reqPath, 'subtitle');
-    if (subtitleId != null) {
-      if (method == 'PUT') {
-        final String suffix =
-            _decodeHeaderValue(request, 'x-hibiki-subtitle-suffix') ?? '';
-        final Directory tmpDir =
-            Directory.systemTemp.createTempSync('hibiki_subtitle_in');
-        final File tmp = File(p.join(tmpDir.path, 'upload.bin'));
-        final IOSink sink = tmp.openWrite();
-        try {
-          await request.read().forEach(sink.add);
-          await sink.close();
-          await svc.importVideoSubtitle(tmp, id: subtitleId, suffix: suffix);
-          return shelf.Response(200);
-        } on ArgumentError catch (e) {
-          return shelf.Response(400, body: 'Invalid subtitle upload: $e');
-        } on StateError {
-          return shelf.Response.notFound('Video not found');
-        } catch (e) {
-          return shelf.Response(500, body: 'Subtitle import failed: $e');
-        } finally {
-          try {
-            await sink.close();
-          } catch (_) {
-            // best-effort
-          }
-          try {
-            tmpDir.deleteSync(recursive: true);
-          } catch (_) {
-            // best-effort
-          }
-        }
-      }
-      if (method != 'GET') return shelf.Response(405);
-      final int episodeIndex = _episodeIndexFromRequest(request);
-      final String? embeddedIndexText =
-          request.url.queryParameters['embeddedStreamIndex'];
-      final File? sub = embeddedIndexText == null
-          ? await svc.resolveVideoSubtitle(subtitleId,
-              episodeIndex: episodeIndex)
-          : await _resolveEmbeddedVideoSubtitle(
-              svc,
-              subtitleId,
-              int.tryParse(embeddedIndexText),
-              episodeIndex,
-            );
-      if (sub == null) return shelf.Response.notFound('Subtitle not found');
-      final int length = sub.lengthSync();
-      return shelf.Response.ok(
-        sub.openRead(),
-        headers: <String, String>{
-          'Content-Type': _guessContentType(sub.path),
-          'Content-Length': '$length',
-        },
-      );
-    }
-
-    // GET/PUT /api/library/videos/<id>/position — 跨设备播放断点（TODO-653）
-    // GET 让 client 拉取 host 真相源进度；PUT 让 client 上报本端进度（host 取较新者）。
-    final String? positionId = _extractVideoId(reqPath, 'position');
-    if (positionId != null) {
-      final int episodeIndex = _episodeIndexFromRequest(request);
-      // 先确认该视频 id（含集下标）在 host DB 真实存在，防止任意 id 写脏 prefs。
-      final File? file =
-          await svc.resolveVideoFile(positionId, episodeIndex: episodeIndex);
-      if (file == null) return shelf.Response.notFound('Video not found');
-      switch (method) {
-        case 'GET':
-          final ({int positionMs, int updatedAtMs}) p = await svc
-              .getVideoPosition(positionId, episodeIndex: episodeIndex);
-          return jsonResponse(<String, dynamic>{
-            'positionMs': p.positionMs,
-            'positionUpdatedAtMs': p.updatedAtMs,
-          });
-        case 'PUT':
-          final String body = await request.readAsString();
-          Map<String, dynamic> json;
-          try {
-            json = jsonDecode(body) as Map<String, dynamic>;
-          } catch (_) {
-            return shelf.Response(400, body: 'Invalid JSON');
-          }
-          final int posMs = (json['positionMs'] as num?)?.toInt() ?? 0;
-          final int updatedAtMs =
-              (json['positionUpdatedAtMs'] as num?)?.toInt() ?? 0;
-          await svc.putVideoPosition(positionId, posMs, updatedAtMs,
-              episodeIndex: episodeIndex);
-          return shelf.Response(200);
-        default:
-          return shelf.Response(405);
-      }
-    }
-
-    // GET/PUT /api/library/videos/<id>/playback — 播放偏好跨设备同步（BUG-1620
-    // 调轴起步，播放偏好同步泛化批扩展为统一带戳字段模型：调轴/音轨/副字幕源/
-    // 副字幕调轴；与 /position 分支对称）。GET 拉 host 生效状态；PUT 上报本端
-    // 带戳字段（host 侧逐字段「严格较新时间戳者胜」+ clamp + 存在性闸门，见
-    // [VideoPlaybackSyncHost]）。老 host 无此分支 → 404，client 上报 best-effort
-    // 吞掉即可（本地 prefs 已持久化）。
-    final String? playbackId = _extractVideoId(reqPath, 'playback');
-    if (playbackId != null) {
-      // 显式 `as`：[VideoPlaybackSyncHost] 不是 [FushiLibraryHostService] 的子
-      // 类型，Dart 不做交集提升（与 [VideoDeletionHost] 的探测写法一致）。
-      if (svc is! VideoPlaybackSyncHost) {
-        return shelf.Response.notFound('Video playback sync not supported');
-      }
-      final VideoPlaybackSyncHost playbackHost = svc as VideoPlaybackSyncHost;
-      // 存在性闸门（一次 DB 单行查询）：防任意 id 写脏 prefs / 枚举探测。
-      if (!await svc.videoExists(playbackId)) {
-        return shelf.Response.notFound('Video not found');
-      }
-      switch (method) {
-        case 'GET':
-          final VideoPlaybackSyncState s =
-              await playbackHost.getVideoPlayback(playbackId);
-          return jsonResponse(s.toJson());
-        case 'PUT':
-          final String body = await request.readAsString();
-          Map<String, dynamic> json;
-          try {
-            json = jsonDecode(body) as Map<String, dynamic>;
-          } catch (_) {
-            return shelf.Response(400, body: 'Invalid JSON');
-          }
-          await playbackHost.putVideoPlayback(
-              playbackId, VideoPlaybackSyncState.fromJson(json));
-          return shelf.Response(200);
-        default:
-          return shelf.Response(405);
-      }
-    }
-
-    // GET /api/library/videos/<id>/clipaudio?startMs=&endMs=&episode=&audioStreamIndex=
-    //     &audioStreamCount=&ac=&bitrate= — BUG-1004：host 端本地裁 mining 句子音频段并回传。
-    // 需 Basic 鉴权（中间件已处理，clipaudio 不在 /stream 豁免名单）。client ffmpeg 打不开
-    // host 自签 https / token 流（移动端自编 ffmpeg-kit TLS pin 残余缺口）时改走此端点——host
-    // 用本地文件裁、不经网络/TLS，只回传几十 KB 的成品音频，不做远端 seek。老 host 无此分支
-    // 对该子路径的 GET 走下方兜底 → 404，client 据此回退直连 ffmpeg 抽取（Never break userspace）。
-    final String? clipAudioId = _extractVideoId(reqPath, 'clipaudio');
-    if (clipAudioId != null) {
-      if (method != 'GET') return shelf.Response(405);
-      final int episodeIndex = _episodeIndexFromRequest(request);
-      final int? startMs =
-          int.tryParse(request.url.queryParameters['startMs'] ?? '');
-      final int? endMs =
-          int.tryParse(request.url.queryParameters['endMs'] ?? '');
-      if (startMs == null || endMs == null || endMs <= startMs) {
-        return shelf.Response(400, body: 'Invalid clip range');
-      }
-      final int? audioStreamIndex =
-          int.tryParse(request.url.queryParameters['audioStreamIndex'] ?? '');
-      final int? audioStreamCount =
-          int.tryParse(request.url.queryParameters['audioStreamCount'] ?? '');
-      final int audioChannels =
-          int.tryParse(request.url.queryParameters['ac'] ?? '') ?? 1;
-      final String audioBitrate =
-          request.url.queryParameters['bitrate'] ?? '64k';
-      final File? clip = await svc.clipVideoAudio(
-        clipAudioId,
-        startMs: startMs,
-        endMs: endMs,
-        episodeIndex: episodeIndex,
-        audioStreamIndex: audioStreamIndex,
-        audioStreamCount: audioStreamCount,
-        audioChannels: audioChannels,
-        audioBitrate: audioBitrate,
-      );
-      if (clip == null) {
-        return shelf.Response.notFound('Clip unavailable');
-      }
-      try {
-        final Uint8List bytes = await clip.readAsBytes();
-        return shelf.Response.ok(
-          bytes,
-          headers: <String, String>{'Content-Type': 'audio/aac'},
-        );
-      } finally {
-        // 裁到独立临时目录，回传后连目录一并清（clipVideoAudio 建的 temp 目录）。
-        try {
-          clip.parent.deleteSync(recursive: true);
-        } catch (_) {
-          // best-effort
-        }
-      }
-    }
-
-    // PUT /api/library/videos/<id> — client→host 上传本地视频文件并注册进 host 视频库
-    // （syncVideoFiles 开关驱动的 live push）。走到此处的 PUT 必是「裸 id 无 suffix」：
-    // 所有带 suffix 的端点（cover/streamurl/stream/subtitle/position）已在上方消化（非
-    // GET 的 suffix 请求返回 405，position 的 PUT 已被上面 switch 接管）。id 允许含 `/`
-    // （bookUid 形如 video/xxx），但拒 `..` / `\`（路径穿越）。title / 原始文件名经
-    // URL-encode 走 header（HTTP header 只收 ASCII，日文标题必须编码）。
-    if (method == 'PUT' && reqPath.startsWith('/api/library/videos/')) {
-      final String? id = _extractBareVideoId(reqPath);
-      if (id == null) {
-        return shelf.Response(400, body: 'Invalid video id');
-      }
-      final String title =
-          _decodeHeaderValue(request, 'x-hibiki-video-title') ?? id;
-      final String? fileName =
-          _decodeHeaderValue(request, 'x-hibiki-video-filename');
-      final Directory tmpDir =
-          Directory.systemTemp.createTempSync('hibiki_video_in');
-      final File tmp = File(p.join(tmpDir.path, 'upload.bin'));
-      final IOSink sink = tmp.openWrite();
-      try {
-        await request.read().forEach(sink.add);
-        await sink.close();
-        await svc.importVideo(tmp,
-            id: id, title: title, originalFileName: fileName);
-        return shelf.Response(200);
-      } catch (e) {
-        try {
-          await sink.close();
-        } catch (_) {
-          // best-effort
-        }
-        return shelf.Response(500, body: 'Video import failed: $e');
-      } finally {
-        try {
-          tmpDir.deleteSync(recursive: true);
-        } catch (_) {
-          // best-effort
-        }
-      }
-    }
-
-    // DELETE /api/library/videos/<id> — client→host 删除远端视频。两个来源：远端视频卡
-    // 长按「删除」，以及本机删除时选了「从所有设备删除」后同步把墓碑推给 host。
-    // 与 PUT 同样只接「裸 id 无 suffix」（带 suffix 的端点已在上方消化）。
-    //
-    // host 未实现 [VideoDeletionHost]（旧版本 app / 测试 fake）时**不接管**，落到下方
-    // 404 —— 这正是 client 侧的能力探测信号，[InterconnectSyncBackend.deleteRemoteVideo]
-    // 按 404/405 判「该 host 不支持」并优雅降级，不报错给用户。
-    // 204 与其它资产链的 DELETE（[_serveAssetPackage]）保持同一成功码。
-    if (method == 'DELETE' && reqPath.startsWith('/api/library/videos/')) {
-      final String? id = _extractBareVideoId(reqPath);
-      if (id == null) {
-        return shelf.Response(400, body: 'Invalid video id');
-      }
-      // 显式 `as` 而不是靠类型提升：[VideoDeletionHost] 不是 [FushiLibraryHostService]
-      // 的子类型，Dart 不做交集提升（写 `svc.deleteVideo` 会报未定义）。与
-      // [DeletionTombstoneHost] 的探测写法一致。
-      if (svc is! VideoDeletionHost) {
-        return shelf.Response.notFound('Video deletion not supported');
-      }
-      try {
-        await (svc as VideoDeletionHost).deleteVideo(id);
-        return shelf.Response(204);
-      } catch (e) {
-        return shelf.Response(500, body: 'Video delete failed: $e');
-      }
-    }
-
-    return shelf.Response.notFound('Not found');
-  }
-
-  /// 从请求 header 读一个 URL-encoded 值并解码（HTTP header 只收 ASCII，非 ASCII 值
-  /// 由 client 端 [Uri.encodeComponent] 编码）。缺失/空返回 null；解码失败退回原文。
-  static String? _decodeHeaderValue(shelf.Request request, String name) {
-    final String? raw = request.headers[name];
-    if (raw == null || raw.isEmpty) return null;
-    try {
-      return Uri.decodeComponent(raw);
-    } catch (_) {
-      return raw;
-    }
-  }
-
-  Map<String, Object?> _remoteVideoJsonForRequest(
-    RemoteVideoInfo video,
-    shelf.Request request,
-  ) {
-    final Map<String, Object?> json = video.toJson()
-      ..remove('coverUrl')
-      ..remove('hasCover');
-    if (_coverFile(video.coverPath) != null) {
-      final String encodedId = Uri.encodeFull(video.id);
-      json['hasCover'] = true;
-      json['coverUrl'] = request.requestedUri.replace(
-        path: '/api/library/videos/$encodedId/cover',
-        queryParameters: <String, String>{},
-      ).toString();
-    }
-    return json;
-  }
-
-  Future<List<RemoteVideoEmbeddedSubtitleTrack>>
-      _embeddedSubtitleTracksForRequest(
-    File videoFile,
-    shelf.Request request,
-    String videoId,
-    int episodeIndex,
-  ) async {
-    final List<EmbeddedSubtitleTrack> tracks =
-        await listEmbeddedSubtitleTracks(videoFile.path);
-    final String encodedId = Uri.encodeFull(videoId);
-    final String videoStem = p.basenameWithoutExtension(videoFile.path);
-    return <RemoteVideoEmbeddedSubtitleTrack>[
-      for (final EmbeddedSubtitleTrack track in tracks)
-        _remoteEmbeddedSubtitleTrackForRequest(
-          track,
-          request,
-          encodedId,
-          videoStem,
-          episodeIndex,
-        ),
-    ];
-  }
-
-  RemoteVideoEmbeddedSubtitleTrack _remoteEmbeddedSubtitleTrackForRequest(
-    EmbeddedSubtitleTrack track,
-    shelf.Request request,
-    String encodedId,
-    String videoStem,
-    int episodeIndex,
-  ) {
-    final String? extension = subtitleExtensionForCodec(track.codec);
-    final bool isText = extension != null;
-    return RemoteVideoEmbeddedSubtitleTrack(
-      streamIndex: track.streamIndex,
-      codec: track.codec,
-      language: track.language,
-      title: track.title,
-      isText: isText,
-      url: isText
-          ? request.requestedUri.replace(
-              path: '/api/library/videos/$encodedId/subtitle',
-              queryParameters: <String, String>{
-                'embeddedStreamIndex': '${track.streamIndex}',
-                if (episodeIndex > 0) 'episode': '$episodeIndex',
-              },
-            ).toString()
-          : null,
-      fileName: isText
-          ? '${_safeDownloadStem(videoStem)}.embedded.${track.streamIndex}$extension'
-          : null,
-    );
-  }
-
-  static String _safeDownloadStem(String value) {
-    final String safe = value.replaceAll(RegExp(r'[^A-Za-z0-9._-]+'), '_');
-    return safe.isEmpty ? 'video' : safe;
-  }
-
-  Future<File?> _resolveEmbeddedVideoSubtitle(
-    FushiLibraryHostService service,
-    String id,
-    int? streamIndex,
-    int episodeIndex,
-  ) async {
-    if (streamIndex == null || streamIndex < 0) return null;
-    final File? videoFile =
-        await service.resolveVideoFile(id, episodeIndex: episodeIndex);
-    if (videoFile == null) return null;
-    final List<EmbeddedSubtitleTrack> tracks =
-        await listEmbeddedSubtitleTracks(videoFile.path);
-    for (final EmbeddedSubtitleTrack track in tracks) {
-      if (track.streamIndex != streamIndex) continue;
-      if (subtitleFormatForCodec(track.codec) == null) return null;
-      return extractEmbeddedSubtitleTrackFile(
-        videoPath: videoFile.path,
-        streamIndex: track.streamIndex,
-        codec: track.codec,
-      );
-    }
-    return null;
-  }
-
-  Future<File?> _resolveVideoCover(
-    FushiLibraryHostService service,
-    String id,
-  ) async {
-    // 单行直查（videoCoverPath = 1 次 DB 查询 + stat）。旧实现每张封面请求重跑
-    // 整份 listVideos()（每行一次目录扫描 + 多次 DB 查询），N 张封面 = O(N²)，
-    // 500 视频的封面墙一次浏览拖成分钟级——这是「互联视频极慢」的主根因。
-    return _coverFile(await service.videoCoverPath(id));
-  }
-
-  static File? _coverFile(String? path) {
-    if (path == null || path.isEmpty) return null;
-    final File file = File(path);
-    return file.existsSync() ? file : null;
-  }
-
-  String _generateVideoToken() {
-    final Random random = Random.secure();
-    final List<int> bytes = List<int>.generate(24, (_) => random.nextInt(256));
-    return base64UrlEncode(bytes);
-  }
-
-  void _pruneVideoTokens() {
-    // 视频播放时间长，token 有效期设为 6 小时
-    final DateTime cutoff = _now().subtract(const Duration(hours: 6));
-    _videoStreamTokens.removeWhere(
-      (String _, _VideoStreamToken token) => token.createdAt.isBefore(cutoff),
-    );
-  }
-
-  /// BUG-1568：守住视频流 token 上限。TTL prune 之后仍达到 [_maxVideoStreamTokens]
-  /// 时，按 createdAt 淘汰最旧者直到回到上限内（对照 [_enforceAudioTokenCap]）。
-  /// 签发前调用，使插入新 token 后总数 <= [_maxVideoStreamTokens]。
-  void _enforceVideoTokenCap() {
-    while (_videoStreamTokens.length >= _maxVideoStreamTokens) {
-      String? oldestKey;
-      DateTime? oldestAt;
-      for (final MapEntry<String, _VideoStreamToken> e
-          in _videoStreamTokens.entries) {
-        if (oldestAt == null || e.value.createdAt.isBefore(oldestAt)) {
-          oldestAt = e.value.createdAt;
-          oldestKey = e.key;
-        }
-      }
-      if (oldestKey == null) break;
-      _videoStreamTokens.remove(oldestKey);
-    }
-  }
-
   /// BUG-1568 测试钩子：当前驻留的视频流 token 数（验证签发侧 cap 逐出行为）。
   @visibleForTesting
   int get videoStreamTokenCount => _videoStreamTokens.length;
 
-  /// 聚合（统计 + 收藏）跨设备 live 端点（TODO-1056 phase C）。
-  ///
-  /// GET /api/library/aggregate — materialize host 自己的聚合快照（四张统计表 +
-  /// 挖掘计数 + 收藏词 + 收藏句）返回 JSON，供 client 拉取 host 真相源。
-  /// PUT /api/library/aggregate — client 上报（已在 client 端与 host 并集合并的）
-  /// 快照，host 用 MAX / 并集 upsert 折叠进自己 DB（幂等，删除不跨端传播）。
-  ///
-  /// 鉴权：不在中间件豁免名单，故自动走 Basic token 全量校验（已配对 peer 才可访问）。
-  /// 容错：坏 JSON → 400；service 未接线 → 404；apply 内部错误经 500 暴露（不静默）。
-  Future<shelf.Response> _handleLibraryAggregate(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    switch (method) {
-      case 'GET':
-        final AggregateSnapshot snapshot = await svc.getAggregateSnapshot();
-        return shelf.Response.ok(
-          jsonEncode(snapshot.toJson()),
-          // charset=utf-8 必带：快照里的书名/标题/义项含 CJK，client 用
-          // package:http `.body` 默认按 latin1 解码会乱码（同 _jsonResponse）。
-          headers: <String, String>{
-            'Content-Type': 'application/json; charset=utf-8'
-          },
-        );
-      case 'PUT':
-        final Map<String, dynamic>? json = await readJsonObjectBody(request);
-        if (json == null) return shelf.Response(400, body: 'Invalid JSON');
-        // fromJson 容错：未知高版本 / 缺字段 / 坏行降级为空或跳过，绝不抛。
-        final AggregateSnapshot snapshot = AggregateSnapshot.fromJson(json);
-        await svc.applyAggregateSnapshot(snapshot);
-        return shelf.Response(200);
-      default:
-        return shelf.Response(405);
-    }
-  }
-
-  /// 合集清单跨设备 live 端点（多端库联合视图 §2.3 任务5.2）。
-  ///
-  /// GET /api/library/collections — 返回 host 当前合集全量快照清单
-  /// （[FushiLibraryHostService.getCollectionManifest] 的 canonicalJson），供 client
-  /// 拉取 host 合集真相源做读-合并-写。
-  /// POST /api/library/collections — client 上报本端合集清单，host 经
-  /// [FushiLibraryHostService.mergeCollectionManifest] 并入自己 DB（成员并集 + 移出/
-  /// 删除墓碑防复活 + 手动序整合集 LWW，语义在 CollectionSyncEngine），返回合并后清单。
-  ///
-  /// 鉴权：不在中间件豁免名单，故自动走 Basic token 全量校验（已配对 peer 才可访问，
-  /// 与 aggregate 端点同纪律）。容错：坏 JSON / 非法/高版本清单 → 400（不静默按旧
-  /// 语义误读新字段污染 host 数据，与 [CollectionManifest.fromJson] 的版本闸门一致）。
-  Future<shelf.Response> _handleLibraryCollections(
-    shelf.Request request,
-    String method,
-    String reqPath,
-  ) async {
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-
-    switch (method) {
-      case 'GET':
-        final CollectionManifest manifest = await svc.getCollectionManifest();
-        return shelf.Response.ok(
-          manifest.canonicalJson(),
-          // charset=utf-8 必带：合集名含 CJK，client 用 package:http `.body` 默认按
-          // latin1 解码会乱码（同 aggregate / _jsonResponse）。
-          headers: <String, String>{
-            'Content-Type': 'application/json; charset=utf-8'
-          },
-        );
-      case 'POST':
-        final Map<String, dynamic>? json = await readJsonObjectBody(request);
-        if (json == null) return shelf.Response(400, body: 'Invalid JSON');
-        CollectionManifest incoming;
-        try {
-          incoming = CollectionManifest.fromJson(json);
-        } on FormatException catch (e) {
-          return shelf.Response(400, body: 'Invalid manifest: $e');
-        }
-        final CollectionManifest merged =
-            await svc.mergeCollectionManifest(incoming);
-        return shelf.Response.ok(
-          merged.canonicalJson(),
-          headers: <String, String>{
-            'Content-Type': 'application/json; charset=utf-8'
-          },
-        );
-      default:
-        return shelf.Response(405);
-    }
-  }
-
-  /// Host → paired child service configuration.
-  ///
-  /// This endpoint carries API keys and connection credentials, so its security
-  /// boundary is intentionally narrower than the ordinary library API:
-  /// HTTPS is mandatory, only a per-peer token is accepted, and no write method
-  /// exists. The legacy shared server token remains valid elsewhere for
-  /// compatibility but is explicitly rejected here.
-  Future<shelf.Response> _handleInterconnectServiceConfig(
-    shelf.Request request,
-    String method,
-  ) async {
-    if (method != 'GET') return shelf.Response(405);
-    if (_securityContext == null) {
-      return shelf.Response.forbidden('HTTPS required for service config');
-    }
-    if (!await _validatePeerAuth(request.headers['authorization'])) {
-      return shelf.Response.forbidden('Paired-device token required');
-    }
-    final FushiLibraryHostService? library = _libraryService;
-    if (library is! InterconnectServiceConfigHost) {
-      return shelf.Response.notFound('Service config capability off');
-    }
-    final InterconnectServiceConfigSnapshot snapshot =
-        await (library as InterconnectServiceConfigHost)
-            .getInterconnectServiceConfig();
-    return shelf.Response.ok(
-      jsonEncode(snapshot.toJson()),
-      headers: const <String, String>{
-        'Content-Type': 'application/json; charset=utf-8',
-        'Cache-Control': 'no-store',
-      },
-    );
-  }
-
-  /// GET / PUT `/api/interconnect/profile`：互联「配置文件」（Profile）双向搬运。
-  ///
-  /// 与 `service-config`（自动跟随的小白名单、单向下行）分工不同：这条是用户**显式点
-  /// 一次**的整份 Profile 搬运，双向。安全门与 service-config 同级并更严一档：
-  ///   * **必须 TLS**（`_securityContext != null`）——载荷是用户的整套设置，明文链路上
-  ///     不给（同 BUG-1311 给 service-config 立的规矩）；
-  ///   * **必须已配对 peer token**（`_validatePeerAuth`，未进鉴权豁免名单）；
-  ///   * **host 侧用户开关**：默认关。没有这道门，入站 PUT 就是一条「无 UI 无开关的
-  ///     隐形写入通道」，正是 BUG-988 点名要避免的形状。开关关着返回 403（而不是 404），
-  ///     好让 client 把「host 不支持」与「host 关着」两种情况分开报。
-  ///
-  /// 载荷就是 Profile 的分享 JSON（`.fushiprofile.json` 的内容），凭据剔除与字体路径
-  /// 剥离由 `ProfileRepository.exportProfileToJson` 负责，见
-  /// [InterconnectProfileHost] 的文档。入站一律 `createNew` 导入，绝不覆盖 host 上任何
-  /// 既有 Profile。
-  Future<shelf.Response> _handleInterconnectProfile(
-    shelf.Request request,
-    String method,
-  ) async {
-    if (method != 'GET' && method != 'PUT') return shelf.Response(405);
-    if (_securityContext == null) {
-      return shelf.Response.forbidden('HTTPS required for profile transfer');
-    }
-    if (!await _validatePeerAuth(request.headers['authorization'])) {
-      return shelf.Response.forbidden('Paired-device token required');
-    }
-    final FushiLibraryHostService? library = _libraryService;
-    if (library is! InterconnectProfileHost) {
-      return shelf.Response.notFound('Profile transfer capability off');
-    }
-    final InterconnectProfileHost host = library as InterconnectProfileHost;
-    if (!await host.isInterconnectProfileTransferEnabled()) {
-      return shelf.Response.forbidden('Profile transfer disabled on host');
-    }
-    if (method == 'GET') {
-      final String json = await host.exportInterconnectProfile();
-      return shelf.Response.ok(
-        json,
-        headers: const <String, String>{
-          'Content-Type': 'application/json; charset=utf-8',
-          'Cache-Control': 'no-store',
-        },
-      );
-    }
-    final String body = await request.readAsString();
-    if (body.trim().isEmpty) {
-      return shelf.Response(400, body: 'Empty profile payload');
-    }
-    try {
-      final String name = await host.importInterconnectProfile(body);
-      return jsonResponse(<String, dynamic>{'name': name});
-    } on FormatException catch (e) {
-      // 载荷不是合法的 Profile 导出（魔数/版本/结构不对）：400，且 host 侧 DB 零改动
-      // （解析在写库之前，见 ProfileRepository.parseProfileExport）。
-      return shelf.Response(400, body: 'Invalid profile payload: ${e.message}');
-    }
-  }
-
-  /// GET `/api/tombstones`：列 host 全部删除墓碑为 JSON 数组，供 client 拉取后与本地
-  /// 在库键求交、弹逐条确认删本地（显式确认式删除传播，host→client 消费方向）。走已配对
-  /// peer 的 Basic token 校验（与 collections/aggregate 同纪律，未进鉴权豁免名单）。
-  Future<shelf.Response> _handleTombstones(
-    shelf.Request request,
-    String method,
-  ) async {
-    if (method != 'GET') return shelf.Response(405);
-    final FushiLibraryHostService? svc = _libraryService;
-    if (svc == null) return shelf.Response.notFound('Library service off');
-    // 可选能力探测：host 未实现删除墓碑列举（老 host / 测试 fake）→ 404，client 侧
-    // getRemoteDeletionTombstones 已优雅降级（不崩、跳过删除墓碑消费）。
-    if (svc is! DeletionTombstoneHost) {
-      return shelf.Response.notFound('No deletion tombstone capability');
-    }
-    final List<({String mediaType, String itemKey, int deletedAt})> rows =
-        await (svc as DeletionTombstoneHost).listDeletionTombstones();
-    final List<Map<String, Object?>> body = <Map<String, Object?>>[
-      for (final r in rows)
-        <String, Object?>{
-          'mediaType': r.mediaType,
-          'itemKey': r.itemKey,
-          'deletedAt': r.deletedAt,
-        },
-    ];
-    return shelf.Response.ok(
-      jsonEncode(body),
-      // charset=utf-8 必带：itemKey 可能含 CJK（书名派生 key），client 按 latin1 默认
-      // 解码会乱码（同 collections / aggregate）。
-      headers: <String, String>{
-        'Content-Type': 'application/json; charset=utf-8'
-      },
-    );
-  }
-
   /// BUG-908(a) 测试钩子：当前驻留的音频 token 数（验证 cap 逐出行为）。
   @visibleForTesting
   int get remoteAudioTokenCount => _audioTokens.count;
-
-  /// TODO-961 M1：清掉 [_pairSessionTtl] 之前创建的配对会话。对照
-  /// [RemoteAudioTokenStore.prune] / [_pruneVideoTokens]：按 createdAt + 注入的 [_now] 判定，
-  /// 可单测。在 pair/v2 创建与 confirm 两处调用，使过期会话既不堆积也不可被 confirm。
-  void _prunePairSessions() {
-    final DateTime cutoff = _now().subtract(_pairSessionTtl);
-    _pairSessions.removeWhere(
-      (String _, FushiPairSession s) => s.createdAt.isBefore(cutoff),
-    );
-  }
-
-  /// TODO-961 M1：守住会话上限。prune 之后仍超过 [_maxPairSessions] 时，按 createdAt
-  /// 淘汰最旧的会话直到回到上限内（攻击者用全新 nonce 高频发起 pair/v2、每个都还在
-  /// TTL 内时的兜底）。正常一次一会话（_pairDialogOpen 串行审批），此路径几乎不触发。
-  void _enforcePairSessionCap() {
-    while (_pairSessions.length >= _maxPairSessions) {
-      String? oldestKey;
-      DateTime? oldestAt;
-      for (final MapEntry<String, FushiPairSession> e
-          in _pairSessions.entries) {
-        if (oldestAt == null || e.value.createdAt.isBefore(oldestAt)) {
-          oldestAt = e.value.createdAt;
-          oldestKey = e.key;
-        }
-      }
-      if (oldestKey == null) break;
-      _pairSessions.remove(oldestKey);
-    }
-  }
 
   /// 测试钩子：当前进行中的配对会话数（验证 prune/cap 行为）。
   @visibleForTesting
@@ -2520,167 +584,6 @@ class FushiSyncServer {
   /// TODO-961 M3 测试钩子：当前限速器跟踪的来源记录数（验证 prune 不泄漏）。
   @visibleForTesting
   int get pinRateLimitTrackedSourceCount => _pinRateLimiter.trackedSourceCount;
-
-  Future<shelf.Response> _handlePropfind(
-      shelf.Request request, String davPath, String fsPath) async {
-    final depth = request.headers['depth'] ?? '1';
-    // BUG-908(b)：逐项 stat 一律异步，避免在事件循环上做阻塞式系统调用（大目录
-    // PROPFIND 会串起成百上千次同步 stat，卡住整个 server）。
-    final entity = await FileSystemEntity.type(fsPath);
-
-    if (entity == FileSystemEntityType.notFound) {
-      return shelf.Response.notFound('Not found');
-    }
-
-    final entries = <_DavEntry>[];
-    final normPath = davPath.endsWith('/') ? davPath : '$davPath/';
-
-    if (entity == FileSystemEntityType.directory) {
-      entries.add(_DavEntry(
-        href: normPath,
-        isCollection: true,
-        displayName: p.basename(fsPath),
-        contentLength: 0,
-      ));
-
-      if (depth == '1') {
-        final dir = Directory(fsPath);
-        await for (final child in dir.list()) {
-          final childName = p.basename(child.path);
-          final isDir = child is Directory;
-          final childHref = '$normPath$childName${isDir ? '/' : ''}';
-          // BUG-908(b)：文件长度用异步 stat（await for 循环里安全 await，不打乱 XML
-          // 组装顺序）；目录不必取长度。
-          final length = isDir ? 0 : (await (child as File).stat()).size;
-          entries.add(_DavEntry(
-            href: childHref,
-            isCollection: isDir,
-            displayName: childName,
-            contentLength: length,
-          ));
-        }
-      }
-    } else {
-      final file = File(fsPath);
-      // BUG-908(b)：单文件长度也用异步 stat。
-      final int fileLength = (await file.stat()).size;
-      entries.add(_DavEntry(
-        href: davPath,
-        isCollection: false,
-        displayName: p.basename(fsPath),
-        contentLength: fileLength,
-      ));
-    }
-
-    final xml = StringBuffer('<?xml version="1.0" encoding="utf-8"?>\n')
-      ..write('<d:multistatus xmlns:d="DAV:">\n');
-    for (final entry in entries) {
-      xml
-        ..write('<d:response>\n')
-        ..write('<d:href>${_xmlEscape(Uri.encodeFull(entry.href))}</d:href>\n')
-        ..write('<d:propstat>\n')
-        ..write('<d:prop>\n')
-        ..write(
-            '<d:displayname>${_xmlEscape(entry.displayName)}</d:displayname>\n')
-        ..write('<d:resourcetype>')
-        ..write(entry.isCollection ? '<d:collection/>' : '')
-        ..write('</d:resourcetype>\n');
-      if (!entry.isCollection) {
-        xml.write(
-            '<d:getcontentlength>${entry.contentLength}</d:getcontentlength>\n');
-      }
-      xml
-        ..write('</d:prop>\n')
-        ..write('<d:status>HTTP/1.1 200 OK</d:status>\n')
-        ..write('</d:propstat>\n')
-        ..write('</d:response>\n');
-    }
-    xml.write('</d:multistatus>');
-
-    return shelf.Response(207,
-        body: xml.toString(),
-        headers: {'Content-Type': 'application/xml; charset=utf-8'});
-  }
-
-  Future<shelf.Response> _handleGet(String fsPath) async {
-    final file = File(fsPath);
-    if (!file.existsSync()) return shelf.Response.notFound('Not found');
-    return shelf.Response.ok(
-      file.openRead(),
-      headers: {
-        'Content-Type': _guessContentType(fsPath),
-        'Content-Length': '${file.lengthSync()}',
-      },
-    );
-  }
-
-  Future<shelf.Response> _handlePut(
-      shelf.Request request, String fsPath) async {
-    final parent = Directory(p.dirname(fsPath));
-    if (!parent.existsSync()) parent.createSync(recursive: true);
-    final file = File(fsPath);
-    final existed = file.existsSync();
-    final sink = file.openWrite();
-    try {
-      await request.read().forEach(sink.add);
-      await sink.close();
-    } catch (e) {
-      // The request stream errored mid-body. Close the sink and remove the
-      // truncated file rather than leaving a corrupt file behind a 201/204
-      // response — matching the download paths' cleanup (HBK-AUDIT-029).
-      try {
-        await sink.close();
-      } catch (_) {/* best-effort: failure is non-critical here */}
-      if (file.existsSync()) {
-        try {
-          file.deleteSync();
-        } catch (_) {/* best-effort: failure is non-critical here */}
-      }
-      return shelf.Response(500, body: 'Write failed');
-    }
-    return shelf.Response(existed ? 204 : 201);
-  }
-
-  Future<shelf.Response> _handleMkcol(String fsPath) async {
-    final dir = Directory(fsPath);
-    if (dir.existsSync()) return shelf.Response(405);
-    dir.createSync(recursive: true);
-    return shelf.Response(201);
-  }
-
-  Future<shelf.Response> _handleDelete(String fsPath) async {
-    final type = FileSystemEntity.typeSync(fsPath);
-    if (type == FileSystemEntityType.notFound) {
-      return shelf.Response.notFound('Not found');
-    }
-    if (type == FileSystemEntityType.directory) {
-      await Directory(fsPath).delete(recursive: true);
-    } else {
-      await File(fsPath).delete();
-    }
-    return shelf.Response(204);
-  }
-
-  Future<shelf.Response> _handleHead(String fsPath) async {
-    final file = File(fsPath);
-    if (!file.existsSync()) return shelf.Response.notFound('Not found');
-    return shelf.Response.ok(null, headers: {
-      'Content-Type': _guessContentType(fsPath),
-      'Content-Length': '${file.lengthSync()}',
-    });
-  }
-
-  /// MIME 推断收敛到 hibiki_core 单一映射表 [mimeTypeForFilePath]（命名统一轮 G8）。
-  /// 旧本地 switch 副本缺 `.webp` → webp 封面按 application/octet-stream 下发，
-  /// 对端 WebView 拒绝内联渲染（BUG-1122）；查共享表后随表修复。保留薄 shim 供
-  /// 本文件既有调用方。
-  static String _guessContentType(String filePath) =>
-      mimeTypeForFilePath(filePath);
-
-  static String _xmlEscape(String s) => s
-      .replaceAll('&', '&amp;')
-      .replaceAll('<', '&lt;')
-      .replaceAll('>', '&gt;');
 }
 
 // ── Range 流式传输辅助（P4-1）────────────────────────────────────────────────
@@ -2780,7 +683,7 @@ ByteRange? parseByteRange(String? rangeHeader, int fileLength) {
 /// - 不可满足的 Range → `416 Range Not Satisfiable` + `Content-Range: bytes */total`。
 ///
 /// 所有路径均加 `Accept-Ranges: bytes`（告知客户端支持 Range）。
-/// Content-Type 由 [FushiSyncServer._guessContentType] 按扩展名确定。
+/// Content-Type 由 [_guessContentType] 按扩展名确定。
 /// 响应体全程流式（`file.openRead(start, end+1)`），不把文件读入内存。
 ///
 /// [etag] 非空时：所有响应带 `ETag` 头；且 Range 请求按 RFC 7233 §3.2 校验
@@ -2800,7 +703,7 @@ Future<shelf.Response> serveFileWithRange(
   }
 
   final int fileLength = file.lengthSync();
-  final String contentType = FushiSyncServer._guessContentType(file.path);
+  final String contentType = _guessContentType(file.path);
   String? rangeHeader = request.headers['range'];
   if (etag != null && rangeHeader != null) {
     final String? ifRange = request.headers['if-range'];
